% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAverageMassError.R
\name{calculateAverageMassError}
\alias{calculateAverageMassError}
\title{Calculate Average Mass Error}
\usage{
calculateAverageMassError(
  data,
  commonColumn,
  spectrumCount,
  startIDX,
  endIDX = NULL,
  roundDecimal = NULL
)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{commonColumn}{Column names of data common for each unique sample: character vector}

\item{spectrumCount}{Spectrum count column name: character}

\item{startIDX}{Index of the first data column to averaged: double}

\item{endIDX}{Index of the last data column to averaged: double}

\item{roundDecimal}{Number of decimal places for precision value rounding (Default: NULL): NULL or double}
}
\value{
A data frame grouped by common columns and the average ppm error for each target m/z value
}
\description{
Calculate the average mass error (in ppm) of target m/z from multiple data columns by grouping using its common columns.
}
