% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMassAccuracy.R
\name{calculateMassAccuracy}
\alias{calculateMassAccuracy}
\title{Calculate Mass Accuracy}
\usage{
calculateMassAccuracy(data, mzArray, targetMZ, roundDecimalPlace = NULL)
}
\arguments{
\item{data}{A data frame of the MS spectral data.}

\item{mzArray}{A character string representing the name of the m/z array column.}

\item{targetMZ}{A numeric vector representing the target m/z values in the spectra array.}

\item{roundDecimalPlace}{A numeric value representing the number of decimal places to be used for precision value rounding. (Default: \code{NULL})}
}
\value{
A data frame of the MS spectral data and the calculated mass accuracy of each target m/z value.
}
\description{
Calculate the mass accuracy (in ppm) of each target m/z value in the spectra array.
}
