% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperClass.R
\name{checkFileExtension}
\alias{checkFileExtension}
\title{Check File Extension}
\usage{
checkFileExtension(parameterName, parameterValue, fileExtension)
}
\arguments{
\item{parameterName}{A character string representing the name of the parameter being validated.}

\item{parameterValue}{A character string representing a file path to match to a file extension specified.}

\item{fileExtension}{A character string representing a file extension (e.g., "JSON").}
}
\value{
Invisible TRUE (logical) if the file path is valid and matches the file extension specified; otherwise returns invisible FALSE (logical)
}
\description{
A helper function to validate a character string representing a file path, and check if it matches a specific file extension..
It ensures that the file has a length of 1, is a character type, is not NA or an empty character string, is a valid file at the specified path,
and checks if the file contains the file extension specified.
}
