% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSpectrumCount.R
\name{checkSpectrumCount}
\alias{checkSpectrumCount}
\title{checkSpectrumCount}
\usage{
checkSpectrumCount(data, commonColumn, sampleID, spectrumCount, threshold = 20)
}
\arguments{
\item{data}{Data frame that contains spectral information of samples: data frame}

\item{commonColumn}{A vector of column names that are common across all samples: character vector}

\item{sampleID}{The name of the column that represents sample ID: character}

\item{spectrumCount}{The name of the column that represents spectrum count: character}

\item{threshold}{Accepted threshold percentage (Default: 20): double}
}
\value{
A list containing statistical result generated from the spectrum check (sample size, total and mean of spectrum count, and the threshold value), and two data frames (samples that passed and failed the check)
}
\description{
Check and filter for samples against a specific threshold set on the average spectrum count (mean) of all samples being processed.
}
