% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureChromatogramRegion.R
\name{configureChromatogramRegion}
\alias{configureChromatogramRegion}
\title{Configure Chromatogram Region}
\usage{
configureChromatogramRegion(
  massCalibrationStart = 0,
  massCalibrationEnd,
  analyteStart = massCalibrationEnd,
  analyteEnd = washStart,
  washStart,
  washEnd = NULL
)
}
\arguments{
\item{analyteStart}{A numeric value that represents the starting time point of the analyte region. (Default: Value used for \code{massCalEnd})}

\item{analyteEnd}{A numeric value that represents the end time point of the analyte region. (Default: Value used for \code{washStart})}

\item{washStart}{A numeric value that represents the starting time point of the wash region.}

\item{washEnd}{A numeric value that represents the end time point of the wash region. (Default: \code{NULL})}

\item{massCalStart}{A numeric value that represents the starting time point of the mass calibration region. (Default: \code{0})}

\item{massCalEnd}{A numeric value that represents the end time point of the mass calibration region.}
}
\value{
A list of lists representing the different chromatogram regions of interest and their respective start and end time points.
}
\description{
Configure the start and end time points of the different chromatogram regions of interest:
\itemize{
\item Prewash region
\item Mass calibration region
\item Analyte region
\item Wash region
}
}
\examples{
# Example 1: Using default values for massCalibrationStart, analyteStart, analyteEnd and washEnd
configureChromatogramRegion(massCalibrationEnd = 0.3, washStart = 5)

# Example 2: Specifying values for analyteStart, analyteEnd and washEnd
configureChromatogramRegion(massCalibrationStart = 0, massCalibrationEnd = 0.3, analyteStart = 2, analyteEnd = 4, washStart = 5, washEnd = 6)
}
