% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSpectrum.R
\name{countSpectrum}
\alias{countSpectrum}
\title{countSpectrum}
\usage{
countSpectrum(data, commonColumn, spectrumCount)
}
\arguments{
\item{data}{Data frame that contains spectral information of samples: data frame}

\item{commonColumn}{Column names that are common across all samples: character vector}

\item{spectrumCount}{Name of the column that represents spectrum count: character}
}
\value{
Summarised data frame grouped by common columns and their respective total number of spectral data (spectrum count)
}
\description{
Count and summarise the total number of spectral data (spectrum count) for each sample present within the data frame.
}
