% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayChromatogramRegion.R
\name{displayChromatogramRegion}
\alias{displayChromatogramRegion}
\title{Display Chromatogram Region}
\usage{
displayChromatogramRegion(plot, maxX, maxY, regionOfInterest, label = NULL)
}
\arguments{
\item{plot}{A ggplot object of a chromatogram.}

\item{maxX}{A numeric value representing the maximum value on the x-axis.}

\item{maxY}{A numeric value representing the maximum value on the y-axis.}

\item{regionOfInterest}{A list representing the start and end time points of a chromatogram region. (Options: Use \code{ticq::configureChromatogram()$regionOfInterest})}

\item{label}{A character string representing the chromatogram region label, (Default: \code{NULL})}
}
\value{
A ggplot object with the added vertical lines representing the chromatogram region start and end time points
}
\description{
Display vertical lines representing the start and end time points of a chromatogram region.
}
