% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayMultipleChromatogram.R
\name{displayMultipleChromatogram}
\alias{displayMultipleChromatogram}
\title{Display Multiple Chromatogram}
\usage{
displayMultipleChromatogram(
  data,
  x,
  firstColumnIndex,
  lastColumnIndex = NULL,
  colour = NULL,
  title = "Chromatogram",
  subtitle = NULL,
  caption = NULL,
  xLabel = x,
  yLabel = NULL,
  colourLabel = colour,
  xTickToggle = TRUE,
  facetWrapBy = NULL,
  facetColumn = NULL,
  facetRow = NULL,
  plotColumn = 1,
  plotRow = 1,
  chromatogramRegion = NULL
)
}
\arguments{
\item{data}{A data frame containing plot data: data frame}

\item{x}{Column name representing the continuous data series to be used for x-axis: character}

\item{firstColumnIndex}{Index of the first data column containing the continuous data series to be used for y-axis: numeric}

\item{lastColumnIndex}{Index of the last data column data containing the continuous data series to be used for y-axis (Default: NULL): NULL or numeric}

\item{colour}{Column name representing data to be used for colour grouping (Default: NULL): NULL or character}

\item{title}{Plot title (Default: "Chromatogram"): NULL or character}

\item{subtitle}{Plot subtitle (Default: NULL): NULL or character}

\item{caption}{Plot caption. This will append a caption value to the column name used for the y-axis (Default: Column name used for y): NULL or character}

\item{xLabel}{x-axis label (Default: Column name used for x): NULL or character}

\item{yLabel}{y-axis label (Default: NULL): NULL or character}

\item{colourLabel}{Colour grouping label (Default: Value used for colour): NULL or character}

\item{xTickToggle}{Toggle to display ticks on x-axis (Default: TRUE, Options: TRUE or FALSE): logical}

\item{facetWrapBy}{Column name representing data to be used for facet wrapping (Default: NULL): NULL or character}

\item{facetColumn}{Number of columns to use for facet wrapping (Default: NULL): NULL or numeric}

\item{facetRow}{Number of rows to use for facet wrapping (Default: NULL): NULL or numeric}

\item{plotColumn}{Number of columns to use for plotting (Default: 1): numeric}

\item{plotRow}{Number of rows to use for plotting (Default: 1): numeric}

\item{chromatogramRegion}{A list representing the different chromatogram regions of interest (Default: NULL, Options: ticq::configureChromatogramRegion()): NULL or list}
}
\description{
Display multiple chromatogram data.
}
\details{
Chromatogram region can be configured using ticq::configureChromatogramRegion().
}
