% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayMultipleTimeSeries.R
\name{displayMultipleTimeSeries}
\alias{displayMultipleTimeSeries}
\title{Display Multiple Time Series}
\usage{
displayMultipleTimeSeries(
  data,
  commonColumn,
  x,
  firstColumnIndex,
  lastColumnIndex = NULL,
  colour = NULL,
  shape = NULL,
  title = "Time Series",
  subtitle = NULL,
  caption = NULL,
  xLabel = x,
  yLabel = NULL,
  colourLabel = colour,
  shapeLabel = shape,
  boundary = NULL,
  boundaryValue = 0,
  referenceData = NULL,
  trendlineToggle = TRUE,
  xTickToggle = TRUE,
  facetWrapBy = NULL,
  facetColumn = NULL,
  facetRow = NULL,
  plotColumn = 1,
  plotRow = 1
)
}
\arguments{
\item{data}{A data frame of the time series data to be used for plotting.}

\item{commonColumn}{A character vector representing the names of the common data columns to be used for data grouping.}

\item{x}{A character string representing the name of the data column to be used for the x-axis.}

\item{firstColumnIndex}{A numeric value representing the index of the first data column to be used for y-axis.}

\item{lastColumnIndex}{A numeric value representing the index of the last data column to be used for y-axis. (Default: \code{NULL})}

\item{colour}{A character string representing the name of the data column to be used for colour grouping. (Default: \code{NULL})}

\item{shape}{A character string representing the name of the data column to be used for shape grouping. (Default: \code{NULL})}

\item{title}{A character string representing the plot title. (Default: \code{"Time Series"})}

\item{subtitle}{A character string representing the plot subtitle. (Default: \code{NULL})}

\item{caption}{A character string representing the plot caption. (Default: \code{NULL})}

\item{xLabel}{character string representing the x-axis label. (Default: \code{x})}

\item{yLabel}{A character string representing the y-axis label. (Default: \code{y})}

\item{colourLabel}{A character string representing the colour grouping label. (Default: \code{colour})}

\item{shapeLabel}{A character string representing the shape grouping label. (Default: \code{shape})}

\item{boundary}{A character string representing the boundary analysis type to perform. (Default: \code{NULL}, Options: \code{"value"}, \code{"percentage"}, or \code{"sd"})}

\item{boundaryValue}{A numeric value to be used for "value" and "percentage" type boundary analysis. (Default: \code{0})}

\item{referenceData}{A data frame containing reference data to be used for "sd" type boundary analysis. (Default: \code{NULL})}

\item{trendlineToggle}{A logical value representing the toggle to display the data trendline. (Default: \code{TRUE}, Options: \code{TRUE} or \code{FALSE})}

\item{xTickToggle}{A logical value representing the toggle to display the ticks on the x-axis. (Default: \code{TRUE}, Options: \code{TRUE} or \code{FALSE})}

\item{facetWrapBy}{A character string representing the name of the data column to be used for facet wrapping. (Default: \code{NULL})}

\item{facetColumn}{A numeric value representing the number of columns to be used for facet wrapping. (Default: \code{NULL})}

\item{facetRow}{A numeric value representing the number of rows to be used for face wrapping. (Default: \code{NULL})}

\item{plotColumn}{A numeric value representing the number of columns to be used for plotting (Default: \code{1})}

\item{plotRow}{A numeric value representing the number of rows to be used for plotting (Default: \code{1})}
}
\value{
A data frame of the boundary analysis result if performed; otherwise, this function does not return any value.
It prints the ggplot objects displaying the multiple time series data.
}
\description{
Display multiple time series data and perform boundary analysis to evaluate data performance over time.
\itemize{
\item Value (Evaluates data by +/- value from 0)
\item Percentage (Evaluates data by +/- value from 100)
\item Standard deviation, sd (Evaluates data using 1st and 2nd standard deviations from average mean of the data or reference data).
}
}
\details{
Type of boundary analysis:
\enumerate{
\item "value" - Uses a boundary value set +/- from 0 as the boundary ranges for evaluation.
\item "percentage" - Uses a boundary value set +/- from 100 as the boundary ranges for evaluation.
\item "sd" - Uses 1st and 2nd standard deviation value from the average mean as the boundary ranges.
To specify an external mean and standard deviation value to be used for analysis, provide reference data (Should have the same data columns as the actual data).
}
}
