% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayPCA.R
\name{displayPCA}
\alias{displayPCA}
\title{Display Principal Component Analysis (PCA)}
\usage{
displayPCA(
  data,
  firstColumnIndex,
  lastColumnIndex = NULL,
  scale = TRUE,
  confidence = 95,
  distribution = NULL,
  colour = NULL,
  shape = NULL,
  subtitle = NULL,
  colourLabel = colour,
  shapeLabel = shape,
  screePlotToggle = TRUE,
  scoresPlotToggle = TRUE,
  loadingsPlotToggle = TRUE,
  biPlotToggle = TRUE
)
}
\arguments{
\item{data}{A data frame containing data to perform PCA.}

\item{firstColumnIndex}{A numeric value representing the index of the first data column to perform PCA on.}

\item{lastColumnIndex}{A numeric value representing the index of the last data column to perform PCA on. (Default: \code{NULL})}

\item{scale}{A logical value representing the scaling used for the PCA. (Default: \code{TRUE}, Options: \code{TRUE} or \code{FALSE})}

\item{confidence}{A numeric value representing the data confidence (\%) to be used for identifying the PC threshold.
The value is also used to draw the data distribution and Hotelling Ellipse on scores plot. (Default: \code{95}, Options: \code{95} or \code{99})}

\item{distribution}{A character string representing the data distribution type to be drawn on the scores plot and biplot. (Default: \code{NULL}, Options: \code{"normal"} or \code{"t"})}

\item{colour}{A character string representing the name of the data column to be used for colour grouping. (Default: \code{NULL})}

\item{shape}{A character string representing the name of the data column to be used for shape grouping. (Default: \code{NULL})}

\item{subtitle}{A character string representing the plot subtitle. (Default: \code{NULL})}

\item{colourLabel}{A character string representing the colour grouping label. (Default: Value used for \code{colour})}

\item{shapeLabel}{A character string representing the shape grouping label. (Default: Value used for \code{shape})}

\item{screePlotToggle}{A logical value representing the toggle to display the generated scree plot. (Default: \code{TRUE}, Options: \code{TRUE} or \code{FALSE})}

\item{scoresPlotToggle}{A logical value representing the toggle to display the generated scores plot. (Default: \code{TRUE}, Options: \code{TRUE} or \code{FALSE})}

\item{loadingsPlotToggle}{A logical value representing the toggle to display the generated loadings plot. (Default: \code{TRUE}, Options: \code{TRUE} or \code{FALSE})}

\item{biPlotToggle}{A logical value representing the toggle to display the generated biplot. (Default: \code{TRUE}, Options: \code{TRUE} or \code{FALSE})}
}
\value{
This function does not return any value. It prints the ggplot objects representing the scree plot, scores plot, loadings plot, and/or biplot.
}
\description{
Perform Principle Component Analysis (PCA) and display analysis outcomes (scree plot, scores plot, loadings plot, and/or biplot).
}
