% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayViolinBox.R
\name{displayViolinBox}
\alias{displayViolinBox}
\title{Display Violin/Box Plot}
\usage{
displayViolinBox(
  data,
  x,
  y,
  colour = NULL,
  shape = NULL,
  title = "Violin/Box Plot",
  subtitle = NULL,
  caption = NULL,
  xLabel = x,
  yLabel = y,
  colourLabel = colour,
  shapeLabel = shape,
  boundary = NULL,
  referenceData = NULL,
  violinPlotToggle = TRUE,
  tailTrimToggle = FALSE,
  boxPlotToggle = TRUE,
  outlierShape = NA,
  boxWidth = 1,
  facetWrapBy = NULL,
  facetColumn = NULL,
  facetRow = NULL
)
}
\arguments{
\item{data}{A data frame containing data to be used for plotting.}

\item{x}{A character string representing the name of the data column to be used for the x-axis.}

\item{y}{A character string representing the name of the data column to be used for the y-axis.}

\item{colour}{A character string representing the name of the data column to be used for colour grouping. (Default: \code{NULL})}

\item{shape}{A character string representing the name of the data column to be used for shape grouping. (Default: \code{NULL})}

\item{title}{A character string representing the plot title. (Default: \code{"Violin/Box Plot"})}

\item{subtitle}{A character string representing the plot subtitle. (Default: \code{NULL})}

\item{caption}{A character string representing the plot caption. (Default: \code{NULL})}

\item{xLabel}{A character string representing the x-axis label. (Default: \code{x})}

\item{yLabel}{A character string representing the y-axis label. (Default: \code{y})}

\item{colourLabel}{A character string representing the colour grouping label. (Default: \code{colour})}

\item{shapeLabel}{A character string representing the shape grouping label. (Default: \code{shape})}

\item{boundary}{A character string representing the boundary analysis type to perform. (Default: \code{NULL}, Options: \code{"value"}, \code{"percentage"}, or \code{"sd"})}

\item{referenceData}{A data frame containing reference data to be used for "sd" type boundary analysis. (Default: \code{NULL})}

\item{violinPlotToggle}{A logical value representing the toggle to display a violin plot. (Default: \code{TRUE}, Options: \code{TRUE} or \code{FALSE})}

\item{tailTrimToggle}{A logical value representing the toggle to trim the tails of the violin plot. (Default: \code{TRUE}, Options: \code{TRUE} or \code{FALSE})}

\item{boxPlotToggle}{A logical value representing the toggle to display a box plot. (Default: \code{TRUE}, Options: \code{TRUE} or \code{FALSE})}

\item{outlierShape}{A numeric value representing the shape style to be used to highlight identified outliers. (Default: \code{NA})}

\item{boxWidth}{A numeric value representing the width size to be used for the box plot. (Default: \code{1})}

\item{facetWrapBy}{A character string representing the name of the data column to be used for facet wrapping. (Default: \code{NULL})}

\item{facetColumn}{A numeric value representing the number of columns to be used for facet wrapping. (Default: \code{NULL})}

\item{facetRow}{A numeric value representing the number of rows to be used for face wrapping. (Default: \code{NULL})}
}
\value{
A data frame of the boundary analysis result if performed; otherwise, this function does not return any value.
It prints the ggplot object displaying violin/box plot.
}
\description{
Display data distribution using violin and/or box plot and perform boundary analysis to evaluate data performance.
\itemize{
\item Value (Evaluates data by +/- value from 0)
\item Percentage (Evaluates data by +/- value from 100)
\item Standard deviation, sd (Evaluates data using 1st and 2nd standard deviations from average mean of the data or reference data)
}
}
\details{
Type of boundary analysis:
\enumerate{
\item "value" - Uses a boundary value set +/- from 0 as the boundary ranges for evaluation.
\item "percentage" - Uses a boundary value set +/- from 100 as the boundary ranges for evaluation.
\item "sd" - Uses 1st and 2nd standard deviation value from the average mean as the boundary ranges.
To specify an external mean and standard deviation value to be used for analysis, provide reference data (Should have the same data columns as the actual data).
}
}
