% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCohort.R
\name{extractCohort}
\alias{extractCohort}
\title{Extract Cohort}
\usage{
extractCohort(input)
}
\arguments{
\item{input}{ANPC sample file name or path: character}
}
\value{
Cohort name
}
\description{
Extract cohort name.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
