% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInstrument.R
\name{extractInstrument}
\alias{extractInstrument}
\title{Extract Instrument}
\usage{
extractInstrument(input)
}
\arguments{
\item{input}{A character vector representing ANPC sample file names or paths.}
}
\value{
A character vector representing instrument names.
}
\description{
Extract instrument name.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
