% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMassError.R
\name{extractMassError}
\alias{extractMassError}
\title{Extract Mass Error}
\usage{
extractMassError(data, mzArray, targetList)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{mzArray}{m/z array column name: character}

\item{targetList}{A list of unique target m/z (Must be equal length as values stored within m/z array): numeric vector}

\item{commonColumn}{Column names of data common for each unique sample: character vector}
}
\value{
A data frame with appended columns representing the calculated mass error (ppm) of each target m/z
}
\description{
Calculate and extract the mass error (in ppm) of each target m/z data detected within the spectra.
}
\details{
The length of the target list containing unique m/z values MUST BE EQUAL to the length of m/z array for each spectra.
}
