% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMatrix.R
\name{extractMatrix}
\alias{extractMatrix}
\title{Extract Matrix}
\usage{
extractMatrix(input)
}
\arguments{
\item{input}{ANPC sample file name or path: character}
}
\value{
Matrix type
}
\description{
Extract matrix type.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
