% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{extractMetadata}
\alias{extractMetadata}
\title{Extract Metadata}
\usage{
extractMetadata(input)
}
\arguments{
\item{input}{A character vector representing ANPC-specific data file names or paths.}
}
\value{
A list of character vectors of the associated metadata.
}
\description{
Extract metadata from Australian National Phenome Centre (ANPC)-specific data files including:
\itemize{
\item Project name
\item Cohort name
\item Matrix type
\item Sample type
\item Method name
\item Instrument name
\item Plate number
}
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
\examples{
extractMetadata(input = "covid19_heidelberg_SER_MS-AA_PAI05_COVp88_261121_QC04_29.json")
}
