% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMethod.R
\name{extractMethod}
\alias{extractMethod}
\title{Extract Method}
\usage{
extractMethod(input)
}
\arguments{
\item{input}{ANPC sample file name or path: character}
}
\value{
Method name
}
\description{
Extract method name.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
