% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPlate.R
\name{extractPlate}
\alias{extractPlate}
\title{Extract Plate}
\usage{
extractPlate(input)
}
\arguments{
\item{input}{ANPC sample file name or path: character}
}
\value{
Plate number
}
\description{
Extract plate number.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
