% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractProject.R
\name{extractProject}
\alias{extractProject}
\title{Extract Project}
\usage{
extractProject(input)
}
\arguments{
\item{input}{A character vector representing ANPC-specific data file names or paths.}
}
\value{
A character vector of the associated project names.
}
\description{
Extract project name.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
