% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSampleType.R
\name{extractSampleType}
\alias{extractSampleType}
\title{Extract Sample Type}
\usage{
extractSampleType(input, matrix)
}
\arguments{
\item{input}{A character vector representing ANPC-specific data file names or paths.}

\item{matrix}{A character vector representing associated matrix types.}
}
\value{
A character vector of the associated sample types.
}
\description{
Extract sample type.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
