% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSpectrum.R
\name{extractSpectrum}
\alias{extractSpectrum}
\title{Extract Spectrum}
\usage{
extractSpectrum(input, inputFiles)
}
\arguments{
\item{input}{Input path (Either a directory path containing JSON files or a single file path): character}

\item{inputFiles}{JSON input files to extract spectral data from: character vector}
}
\value{
A list containing two data frames (passed and failed data)
}
\description{
Extract spectral data from JSON files generated using ExfilMS (https://github.com/vmalnathnambiar/exfilms). Also includes retrievable metadata from sample name or file path.
}
\details{
Metadata extraction only applicable to data acquired at the Australian National Phenome Centre (ANPC).

To evaluate target m/z with TICQ, perform targeted spectra filtering using ExfilMS. The target file used for ExfilMS should match the target file defined in TICQ for comparison
}
