% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTargetMZ.R
\name{extractTargetMZ}
\alias{extractTargetMZ}
\title{extractTargetMZ}
\usage{
extractTargetMZ(
  targetFile = NULL,
  anpcMethodLibrary = NULL,
  roundDecimal = NULL
)
}
\arguments{
\item{targetFile}{TSV document that contains target m/z data (Default: NULL, Options: Published to web URL link or local file path): NULL or character}

\item{anpcMethodLibrary}{ANPC method library (Default: NULL, Options: "MS-AA-POS", "MS-HIL-POS", "MS-HIL-NEG", "MS-RP-POS" or "MS-RP-NEG"): NULL or character}

\item{roundDecimal}{The number of decimal places to round the target m/z values to for analysis (Default: NULL): NULL or double}
}
\value{
A data frame containing the extracted target m/z data
}
\description{
Extract target m/z data from target file (published to web URL or local TSV document) or in-house method library used at the Australian National Phenome Centre (ANPC).
}
\examples{
# Example 1: Specifying a target file with no rounding of decimal places for the target m/z values
targetMZ <- ticq::extractTargetMZ(targetFile = "https://docs.google.com/spreadsheets/d/e/2PACX-1vSeo31hlruA3QuwoESz5IDJ9Nu6ndSAgLTRn3uc45rOPO4BlksfHzh9xtNB22Oes9JOxhEbI4NK-zxl/pub?gid=0&single=true&output=tsv", 
                                  anpcMethodLibrary = NULL, 
                                  roundDecimal = NULL)

# Example 2: Choosing an in-house method library used at the Australian National Phenome Centre (ANPC) and rounding the target m/z values to 4 decimal places
targetMZ <- ticq::extractTargetMZ(targetFile = NULL, 
                                  anpcMethodLibrary = "MS-AA-POS", 
                                  roundDecimal = 4)
}
