% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{formatDirectoryPath}
\alias{formatDirectoryPath}
\title{Format Directory Path}
\usage{
formatDirectoryPath(name, value)
}
\arguments{
\item{name}{A character string representing the name of the object being validated.}

\item{value}{A character string representing a directory path.}
}
\value{
A character string of the validated directory path with a forward slash (/); otherwise, it stops with an error message.
}
\description{
Validate that a character string representing a directory path ends with a forward slash (/).
}
