% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateStat.R
\name{generateStat}
\alias{generateStat}
\title{generateStat}
\usage{
generateStat(data)
}
\arguments{
\item{data}{Continuous data series of numerical values without NA (At least 2 data points): numerical vector}
}
\value{
Data frame containing all associated statistical information
}
\description{
Generate statistical information based on data.
}
\details{
Statistical information includes:
\itemize{
\item Sample size (n)
\item Minimum (min)
\item Mean (mean)
\item Maximum (max)
\item 1st Quartile (Q1)
\item Median
\item 3rd Quartile (Q3)
\item Inter-quantile range (IQR)
\item IQR inner fence (Lower and upper)
\item IQR outer fence (Lower and upper)
\item 1st standard deviation (sd)
\item 2nd standard deviation (sd2)
\item Standard error (se)
\item Margin error (me) - 95\% and 99\%
\item Confidence interval (CI) - 95\% and 99\% (Lower and upper)
\item Confidence quantile range (CQ) - 95\% and 99\% (Lower and upper)
}
}
\examples{
stat <- ticq::generateStat(data = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
print(stat)
}
