% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateStatistic.R
\name{generateStatistic}
\alias{generateStatistic}
\title{Generate Statistic}
\usage{
generateStatistic(data)
}
\arguments{
\item{data}{A numeric vector representing values to be used to generate statistical information.}
}
\value{
A list of statistical information.
}
\description{
Generate statistical information including:
\itemize{
\item Sample size (n)
\item Sum (sum)
\item Minimum (min)
\item Mean (mean)
\item Maximum (max)
\item 1st Quartile (Q1)
\item Median
\item 3rd Quartile (Q3)
\item Inter-quantile range (IQR)
\item IQR inner fence (Lower and upper)
\item IQR outer fence (Lower and upper)
\item Standard deviation (sd)
\item Standard error (se)
\item Margin error (me) - 95\% and 99\%
\item Confidence interval (CI) - 95\% and 99\% (Lower and upper)
\item Confidence quantile range (CQ) - 95\% and 99\% (Lower and upper)
}
}
\examples{
generateStatistic(data = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
}
