% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackage.R
\name{loadPackage}
\alias{loadPackage}
\title{Load Package}
\usage{
loadPackage(packageName)
}
\arguments{
\item{packageName}{A character vector representing the package names to be installed and loaded.}
}
\value{
A logical vector indicating whether each package specified was successfully loaded.
}
\description{
Identify, install, and load packages to use.
}
\examples{
# Example 1: Loading a single package
loadPackage(packageName = "DT")

# Example 2: Loading multiple packages
loadPackage(packageName = c("DT", "dplyr"))
}
