% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultipleTimeSeries.R
\name{plotMultipleTimeSeries}
\alias{plotMultipleTimeSeries}
\title{plotMultipleTimeSeries}
\usage{
plotMultipleTimeSeries(
  data,
  commonColumn,
  x,
  startIDX,
  endIDX = NULL,
  colour = NULL,
  shape = NULL,
  caption = NULL,
  labelX = x,
  labelY = NULL,
  colourLegend = colour,
  shapeLegend = shape,
  boundary = NULL,
  boundaryValue = 0,
  xTickToggle = TRUE,
  nPlotCol = 1,
  nPlotRow = 1
)
}
\arguments{
\item{data}{Data frame that contains plot data: data frame}

\item{commonColumn}{Column names that are common across all samples: character vector}

\item{x}{Name of the column that represents the continuous data series for x-axis: character}

\item{startIDX}{Index of the first data column that contains continuous data series for the y-axis: double}

\item{endIDX}{Index of the last data column data that contains continuous data series for the y-axis (Default: NULL): NULL or double}

\item{colour}{Name of the column used for colour grouping (Default: NULL): NULL or character}

\item{shape}{Name of the column used for shape grouping (Default: NULL): NULL or character}

\item{caption}{Plot caption (Default: NULL): NULL or character}

\item{labelX}{x-axis label (Default: Value used for x): NULL or character}

\item{labelY}{y-axis label (Default: NULL): NULL or character}

\item{colourLegend}{Legend label for colour grouping (Default: Value used for colour): NULL or character}

\item{shapeLegend}{Legend label for shape grouping (Default: Value used for shape)}

\item{boundary}{Boundary analysis to perform based on a +/- value from 0 (value) or 100 (percentage), or up to 2 standard deviation (sd) from average mean (Default: NULL, Options: "value", "percentage", "sd"): NULL or character}

\item{boundaryValue}{Value used for boundary analysis. To specify values for "sd" boundary analysis, use list(mean = meanValue, sd = sdValue, sd2 = sd2Value) (Default: 0): double or list}

\item{xTickToggle}{Toggle ticks on x-axis (Default: TRUE, Options: TRUE or FALSE): boolean}

\item{nPlotCol}{Number of plots in a single columns (Default: 1): double}

\item{nPlotRow}{Number of plot in a single row (Default: 1): double}
}
\value{
Data frame containing the boundary analysis result (if specified)
}
\description{
Plot multiple time series data and perform boundary analysis (if specified).
}
