% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{plotPCA}
\usage{
plotPCA(
  data,
  startIDX,
  endIDX = NULL,
  scale = TRUE,
  confidence = 95,
  distribution = NULL,
  colour = NULL,
  shape = NULL,
  subtitle = NULL,
  colourLegend = colour,
  shapeLegend = shape,
  screePlt = FALSE,
  scoresPlt = FALSE,
  loadingsPlt = FALSE,
  biPlt = FALSE
)
}
\arguments{
\item{data}{Data frame that contains columns representing data to perform PCA on. Should be equal sized and not contain NA values: data frame}

\item{startIDX}{Index of the first data column to perform PCA on: double}

\item{endIDX}{Index of the last data column data to perform PCA on (Default: NULL): NULL or double}

\item{scale}{Scaling status for the PCA (Default: TRUE, Options: TRUE or FALSE): boolean}

\item{confidence}{Data confidence \% (Default: 95): double}

\item{distribution}{Data distribution. Distribution boundary is drawn on scores plot and biplot (Default: NULL, Options: "normal" or "t"): NULL or character}

\item{colour}{Name of the column used for colour grouping (Default: NULL): NULL or character}

\item{shape}{Name of the column used for shape grouping (Default: NULL): NULL or character}

\item{subtitle}{Plot subtitle (Default: NULL): NULL or character}

\item{colourLegend}{Legend label for colour grouping (Default: Value used for colour): NULL or character}

\item{shapeLegend}{Legend label for shape grouping (Default: Value used for shape): NULL or character}

\item{screePlt}{Toggle printing of scree plot (Default: FALSE, Options: TRUE or FALSE): boolean}

\item{scoresPlt}{Toggle printing of scores plot (Default: FALSE, Options: TRUE or FALSE): boolean}

\item{loadingsPlt}{Toggle printing of loadings plot (Default: FALSE, Options: TRUE or FALSE): boolean}

\item{biPlt}{Toggle printing of biplot (Default: FALSE, Options: TRUE or FALSE): boolean}
}
\description{
Perform Principle Component Analysis (PCA) and plot all associated plots (scree plot, scores plot, loadings plot and biplot).
}
