% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{plotTimeSeries}
\usage{
plotTimeSeries(
  data,
  x,
  y,
  colour = NULL,
  shape = NULL,
  title = "Time Series",
  subtitle = NULL,
  caption = NULL,
  labelX = x,
  labelY = y,
  colourLegend = colour,
  shapeLegend = shape,
  boundary = NULL,
  boundaryValue = 0,
  xTick = TRUE
)
}
\arguments{
\item{data}{Data frame that contains plot data: data frame}

\item{x}{Name of the column that represents the continuous data series for x-axis: character}

\item{y}{Name of the column that represents the continuous data series for y-axis: character}

\item{colour}{Name of the column used for colour grouping (Default: NULL): NULL or character}

\item{shape}{Name of the column used for shape grouping (Default: NULL): NULL or character}

\item{title}{Plot title (Default: "Time Series"): character}

\item{subtitle}{Plot subtitle (Default: NULL): NULL or character}

\item{caption}{Plot caption (Default: NULL): NULL or character}

\item{labelX}{x-axis label (Default: Value used for x): NULL or character}

\item{labelY}{y-axis label (Default: Value used for y): NULL or character}

\item{colourLegend}{Legend label for colour grouping (Default: Value used for colour): NULL or character}

\item{shapeLegend}{Legend label for shape grouping (Default: Value used for shape)}

\item{boundary}{Boundary analysis to perform based on a +/- value from 0 (value) or 100 (percentage), or up to 2 standard deviation (sd) from average mean (Default: NULL, Options: "value", "percentage", "sd"): NULL or character}

\item{boundaryValue}{Value used for boundary analysis. For "sd" boundary analysis, use list(mean = meanValue, sd = sdValue, sd2 = sd2Value) (Default: 0): double or list}

\item{xTick}{Toggle ticks on x-axis (Default: TRUE, Options: TRUE or FALSE): boolean}
}
\value{
Data frame containing the boundary analysis result (if specified)
}
\description{
Plot time series data and perform boundary analysis (if specified).
}
