% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViolin.R
\name{plotViolin}
\alias{plotViolin}
\title{plotViolin}
\usage{
plotViolin(
  data,
  x,
  y,
  colour = NULL,
  shape = NULL,
  outlierShape = NA,
  title = "Violin Plot",
  subtitle = NULL,
  caption = NULL,
  labelX = x,
  labelY = y,
  colourLegend = colour,
  shapeLegend = shape,
  boundary = NULL,
  boundaryValue = 0,
  violinPlotToggle = TRUE,
  tailTrimToggle = FALSE,
  boxPlotToggle = TRUE,
  boxWidth = 1
)
}
\arguments{
\item{data}{Data frame that contains plot data: data frame}

\item{x}{Name of the column that represents the continuous data series for x-axis: character}

\item{y}{Name of the column that represents the continuous data series for y-axis: character}

\item{colour}{Name of the column used for colour grouping (Default: NULL): NULL or character}

\item{shape}{Name of the column used for shape grouping (Default: NULL): NULL or character}

\item{outlierShape}{Shape used to identify outliers ((Default: NA): NA or double)}

\item{title}{Plot title (Default: "Time Series"): character}

\item{subtitle}{Plot subtitle (Default: NULL): NULL or character}

\item{caption}{Plot caption (Default: NULL): NULL or character}

\item{labelX}{x-axis label (Default: Value used for x): NULL or character}

\item{labelY}{y-axis label (Default: Value used for y): NULL or character}

\item{colourLegend}{Legend label for colour grouping (Default: Value used for colour): NULL or character}

\item{shapeLegend}{Legend label for shape grouping (Default: Value used for shape)}

\item{boundary}{Boundary analysis to perform based on a +/- value from 0 (value) or 100 (percentage), or up to 2 standard deviation (sd) from average mean (Default: NULL, Options: "value", "percentage", "sd"): NULL or character}

\item{boundaryValue}{Value used for boundary analysis. For "sd" boundary analysis, use list(mean = meanValue, sd = sdValue, sd2 = sd2Value) (Default: 0): double or list}

\item{violinPlotToggle}{Toggle printing of violin plot (Default: TRUE, Options: TRUE or FALSE): boolean}

\item{tailTrimToggle}{Toggle to trim the tails of the violin plot. Applicable only if violinPlotToggle is set to TRUE (Default: FALSE, Options: TRUE or FALSE): boolean}

\item{boxPlotToggle}{Toggle printing of box plot (Default: TRUE, Options: TRUE or FALSE): boolean}

\item{boxWidth}{Width of box plot. Applicable if boxPlotToggle is set to TRUE (Default: 1): double}
}
\value{
Data frame containing the boundary analysis result (if specified)
}
\description{
Plot violin plots and perform boundary analysis (if specified).
}
