% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderTICQ.R
\name{renderTICQ}
\alias{renderTICQ}
\title{Render TICQ}
\usage{
renderTICQ(
  inputPath,
  outputDirectoryPath = paste0(getwd(), "/output/"),
  historicalDataPath = NULL,
  targetFilePath = NULL,
  anpcMethodLibrary = NULL,
  chromatogramRegion = NULL,
  roundDecimalPlace = NULL,
  metadataAnonymiser = FALSE
)
}
\arguments{
\item{inputPath}{A character string representing an existing directory or file path. Should not be NA or an empty character string.}

\item{outputDirectoryPath}{A character string representing a directory path to store output files (HTML report and RDS formatted spectral data). (Default: "/home/vimalnathnambiar/git/github/vmalnathnambiar/ticq/output/").}

\item{historicalDataPath}{A character string representing a RDS file containing historical spectral data associated to the current spectral data being assessed. (Default: \code{NULL}; Options: \code{"/path/to/local/file.rds"})}

\item{targetFilePath}{A character string representing an existing TSV target file path. (Default: \code{NULL}; Options: \code{"https://published-to-web-URL&output=tsv"} or \code{"/path/to/local/file.tsv"})}

\item{anpcMethodLibrary}{ANPC method library (Default: NULL, Options: "MS-AA-POS", "MS-HIL-POS", "MS-HIL-NEG", "MS-RP-POS" or "MS-RP-NEG"): NULL or character}

\item{chromatogramRegion}{A list containing the start and end time points of the different chromatogram regions of interest (Default: NULL, Options: ticq::configureChromatogramRegion()): NULL or list}

\item{roundDecimalPlace}{Number of decimal places for precision value rounding (Default: NULL): NULL or numeric}

\item{metadataAnonymiser}{Toggle to anonymise metadata (Default: FALSE, Options: TRUE or FALSE): logical}
}
\value{
Path to the rendered output file.
}
\description{
Generate a HTML report of TICQ assessment performed on MS spectral data acquired from JSON files generated using \href{https://github.com/vmalnathnambiar/exfilms}{ExfilMS}.
}
\details{
\enumerate{
\item Metadata extraction and anonymisation only applicable to data acquired at the Australian National Phenome Centre (ANPC).
\item To perform a targeted assessment of extracted ions (targeted m/z values) with TICQ, spectral data extracted using ExfilMS should have undergone targeted spectra filtering during the extraction process using a target file or an in-house ANPC method library.
The target file or ANPC method library used should then also be specified in TICQ for a comprehensive assessment.
\item To perform comparison with a historical spectral data, the mentioned data should have also undergo the same extraction process as the current spectral data being assessed.
}
}
