% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveFileList.R
\name{retrieveFileList}
\alias{retrieveFileList}
\title{Retrieve File List}
\usage{
retrieveFileList(inputPath, fileExtension = NULL)
}
\arguments{
\item{inputPath}{Input path: character}

\item{fileExtension}{File extension type (Default: NULL): NULL or character}
}
\value{
A list of file names matching the defined file extension type: character or character vector
}
\description{
Retrieve a list of file names matching a defined file extension type (i.e., "JSON", "XML").
}
