% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveFileName.R
\name{retrieveFileName}
\alias{retrieveFileName}
\title{Retrieve File Name}
\usage{
retrieveFileName(input, fileExtension = NULL)
}
\arguments{
\item{input}{Input path containing files of interest: character}

\item{fileExtension}{File extension to filter for (Default: NULL): NULL or character}
}
\value{
File names matching the file extension defined: character vector
}
\description{
Filter and retrieve file names of a specific extension type stored within an input path.
}
