% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceChromatogramRegion.R
\name{spliceChromatogramRegion}
\alias{spliceChromatogramRegion}
\title{Splice Chromatogram Region}
\usage{
spliceChromatogramRegion(
  data,
  commonColumn,
  spectrumCount,
  chromatogramRegion = NULL,
  retentionTime,
  sumColumn
)
}
\arguments{
\item{data}{A data frame of the MS spectral data.}

\item{commonColumn}{A character vector representing the names of the common data columns to be used for data grouping.}

\item{spectrumCount}{A character string representing the name of the spectrum count column.}

\item{chromatogramRegion}{A list representing the chromatogram region data of interests. (Default: \code{NULL}; Options: \code{configureChromatogramRegion()})}

\item{retentionTime}{A character string representing the name of the retention time column.}

\item{sumColumn}{A character string representing the name of the data column with values to be summed.}
}
\value{
A data frame of the MS spectral data summary grouped by their common data columns and the spliced data of each chromatogram region of interest.
}
\description{
Splice MS spectral data of each data group by summing the values of a data column using the chromatogram region data of interests (start and end time points).
}
