% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumEIC.R
\name{sumEIC}
\alias{sumEIC}
\title{Sum Extracted Ion Current (EIC)}
\usage{
sumEIC(
  data,
  commonColumn,
  spectrumCount,
  firstColumnIndex,
  lastColumnIndex = NULL
)
}
\arguments{
\item{data}{A data frame of the MS spectral data.}

\item{commonColumn}{A character vector representing the names of the common data columns to be used for data grouping.}

\item{spectrumCount}{A character string representing the name of the spectrum count column.}

\item{firstColumnIndex}{A numeric value representing the index of the first data column to be summed.}

\item{lastColumnIndex}{A numeric value representing the index of the last data column to be summed. (Default: \code{NULL})}
}
\value{
A data frame of the MS spectral data grouped by their common data columns and the summed intensity of each target m/z value.
}
\description{
Sum the intensity of each target m/z value of each data group.
}
