% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHelper.R
\name{validateCharacterString}
\alias{validateCharacterString}
\title{Validate Character String}
\usage{
validateCharacterString(parameterName, parameterValue)
}
\arguments{
\item{parameterName}{A character string representing the name of the parameter being validated.}

\item{parameterValue}{A character string to be validated.}
}
\value{
Invisible TRUE if the character string is valid; otherwise, it stops with an error message.
}
\description{
A helper function to validate a character string.
It ensures that the string specified has a length of 1, is a character type, and is not NA or an empty character string.
}
