% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{validateCharacterVectorElement}
\alias{validateCharacterVectorElement}
\title{Validate Character Vector Element}
\usage{
validateCharacterVectorElement(name, value)
}
\arguments{
\item{name}{A character string representing the name of the object being validated.}

\item{value}{A character vector.}
}
\value{
Invisible TRUE if the character vector is valid; otherwise, it stops with an error message.
}
\description{
Validate that a vector has a length greater than 0, is a character type, and does not contain any NA or empty string elements.
}
