% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateChromatogramRegion.R
\name{validateChromatogramRegion}
\alias{validateChromatogramRegion}
\title{Validate Chromatogram Region}
\usage{
validateChromatogramRegion(chromatogramRegion)
}
\arguments{
\item{chromatogramRegion}{A list representing the different chromatogram regions of interest (Options: ticq::configureChromatogramRegion()): list}
}
\value{
Validation status (TRUE or FALSE)
}
\description{
Validate chromatogram region data of interests (prewash, mass calibration, analyte and wash).
}
