% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{validateDirectoryFileExist}
\alias{validateDirectoryFileExist}
\title{Validate Directory and File Path}
\usage{
validateDirectoryFileExist(name, value, pathType = "both")
}
\arguments{
\item{name}{A character string representing the name of the object being validated.}

\item{value}{A character string representing either a directory or file path.}

\item{pathType}{A character string representing the path type to check for. (Default: \code{"both"}; Options: \code{"directory"}, \code{"file"}, or \code{"both"})}
}
\value{
Invisible TRUE if the directory or file path exists; otherwise, it stops with an error message.
}
\description{
Validate that a character string representing a directory or file path exists based on a specific path type check (directory, file, or both).
}
