% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHelper.R
\name{validateDirectoryPathFormat}
\alias{validateDirectoryPathFormat}
\title{Validate Directory Path Format}
\usage{
validateDirectoryPathFormat(parameterName, parameterValue)
}
\arguments{
\item{parameterName}{A character string representing the name of the parameter being validated.}

\item{parameterValue}{A character string representing the directory path to be validated.}
}
\value{
A character string representing the validated and standardised directory path with a forward slash (/); otherwise, it stops with an error message.
}
\description{
A helper function to validate and standardize the format of a character string representing a directory path.
It ensures that the directory path specified has a length of 1, is a character type, is not NA or an empty character string,
and ends with a forward slash (/).
}
