% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHelper.R
\name{validateFileExtension}
\alias{validateFileExtension}
\title{Validate File Extension}
\usage{
validateFileExtension(parameterName, parameterValue, fileExtension)
}
\arguments{
\item{parameterName}{A character string representing the name of the parameter being validated.}

\item{parameterValue}{A character string representing a file path to be checked.}

\item{fileExtension}{A character string representing a file extension to check for.}
}
\value{
Invisible TRUE if the file path is valid and matches the file extension specified; otherwise returns invisible FALSE.
}
\description{
A helper function to validate a character string representing a file path if it matches a specific file extension.
It ensures that the file specified has a length of 1, is a character type, is not NA or an empty character string, is a valid file at the specified path,
and validates if the file matches the file extension specified.
}
