% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{validateLogicalValue}
\alias{validateLogicalValue}
\title{Validate Logical Value}
\usage{
validateLogicalValue(name, value)
}
\arguments{
\item{name}{A character string representing the name of the object being validated.}

\item{value}{A logical value.}
}
\value{
Invisible TRUE if the logical value is valid; otherwise, it stops with an error message.
}
\description{
Validate that a value has a length of 1 and is a logical type, and is not NA.
}
