% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{validateMissingNumericValue}
\alias{validateMissingNumericValue}
\title{Validate Missing Numeric Value}
\usage{
validateMissingNumericValue(name, value)
}
\arguments{
\item{name}{A character string representing the name of the object being validated.}

\item{value}{A numeric value.}
}
\value{
Invisible TRUE if the numeric value is valid; otherwise, it stops with an error message.
}
\description{
Validate that a value is either NA (missing) or has a length of 1 and is a numeric type.
}
