% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHelper.R
\name{validateNumericValue}
\alias{validateNumericValue}
\title{Validate Numeric Value}
\usage{
validateNumericValue(parameterName, parameterValue)
}
\arguments{
\item{parameterName}{A character string representing the name of the parameter being validated.}

\item{parameterValue}{A numeric value to be validated.}
}
\value{
Invisible TRUE if the numeric value is valid; otherwise, it stops with an error message.
}
\description{
A helper function to validate a numeric value.
It ensures that the value specified has a length of 1 and is a numeric type.
}
