% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{validateNumericVectorElement}
\alias{validateNumericVectorElement}
\title{Validate Numeric Vector Element}
\usage{
validateNumericVectorElement(name, value)
}
\arguments{
\item{name}{A character string representing the name of the object being validated.}

\item{value}{A numeric vector.}
}
\value{
Invisible TRUE if the numeric vector is valid; otherwise, it stops with an error message.
}
\description{
Validate that a vector has a length greater than 0, is a numeric type, and does not contain any NA elements.
}
