% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{validateRegionOfInterest}
\alias{validateRegionOfInterest}
\title{Validate Region of Interest}
\usage{
validateRegionOfInterest(name, value)
}
\arguments{
\item{name}{A character string representing the name of the object being validated.}

\item{value}{A list representing the start and end time points of a chromatogram region of interest.}
}
\value{
Invisible TRUE if data of the chromatogram region of interest is valid; otherwise, invisible FALSE along with an error message.
}
\description{
Validate that a list representing the time points of a chromatogram region of interest has a length of 2 (start and end) and
that each time point value is either NULL or has a length of 1 and is a numeric type.
}
