% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymiseMetadata.R
\name{anonymiseMetadata}
\alias{anonymiseMetadata}
\title{Anonymise Metadata}
\usage{
anonymiseMetadata(
  data,
  sampleID,
  project,
  cohort,
  projectCohort,
  method,
  instrument,
  plate
)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{sampleID}{Sample ID column name: character}

\item{project}{Project column name: character}

\item{cohort}{Cohort column name: character}

\item{projectCohort}{Project + Cohort column name: character}

\item{method}{Method column name: character}

\item{instrument}{Instrument column name: character}

\item{plate}{Plate number column name: character}
}
\value{
A data frame with anonymised sample metadata
}
\description{
Anonymise sample metadata including sample ID, project name, cohort name, method name, instrument name, and plate number.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
\examples{
data <- data.frame(sampleID = "covid19_heidelberg_SER_MS-AA_PAI05_COVp88_261121_QC04_29", project = "COVID-19", cohort = "Heidelberg",
                   projectCohort = "COVID-19 Heidelberg", method = "MS-AA-POS", instrument = "PAI05", plate = "88")
data <- ticq::anonymiseMetadata(data = data, sampleID = "sampleID", project = "project", cohort = "cohort", projectCohort = "projectCohort",
                                method = "method", instrument = "instrument", plate = "plate")
print(data)
}
