% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymiseMetadata.R
\name{anonymiseMetadata}
\alias{anonymiseMetadata}
\title{anonymiseMetadata}
\usage{
anonymiseMetadata(
  data,
  sampleID,
  project,
  cohort,
  projectCohort,
  method,
  instrument,
  plate
)
}
\arguments{
\item{data}{Data frame that contains spectral information of samples: data frame}

\item{commonColumn}{Column names that are common across all samples: character vector}
}
\value{
Data frame containing spectral information of anonymised samples
}
\description{
Anonymise sample metadata.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
Information anonymised includes sample ID, project name, cohort name, method name, instrument name, and plate number.
}
\examples{
data <- data.frame(sampleID = "covid19_heidelberg_SER_MS-AA_PAI05_COVp88_261121_QC04_29",
                   project = "COVID-19",
                   cohort = "Heidelberg",
                   projectCohort = "COVID-19 Heidelberg",
                   method = "MS-AA-POS",
                   instrument = "PAI05",
                   plate = 88)
data <- ticq::anonymiseMetadata(data = data,
                                sampleID = "sampleID",
                                project = "project",
                                cohort = "cohort",
                                projectCohort = "projectCohort",
                                method = "method",
                                instrument = "instrument",
                                plate = "plate")
print(data)
}
