% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymiseMetadata.R
\name{anonymiseMetadata}
\alias{anonymiseMetadata}
\title{Anonymise Metadata}
\usage{
anonymiseMetadata(
  data,
  sampleID,
  project,
  cohort,
  projectCohort,
  method,
  instrument,
  plate
)
}
\arguments{
\item{data}{A data frame of the MS spectral data.}

\item{sampleID}{A character string representing the name of the sample ID column.}

\item{project}{A character string representing the name of the project column.}

\item{cohort}{A character string representing the name of the cohort column.}

\item{projectCohort}{A character string representing the name of the combined project and cohort column.}

\item{method}{A character string representing the name of the method column.}

\item{instrument}{A character string representing the name of the instrument column.}

\item{plate}{A character string representing the name of the plate column.}
}
\value{
A data frame of the MS spectral data and its associated anonymised metadata in respective data columns.
}
\description{
Anonymise metadata associated with Australian National Phenome Centre (ANPC)-specific data files including:
\itemize{
\item Sample ID
\item Project name
\item Cohort name
\item Method name
\item Instrument name
\item Plate number
}
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
\examples{
data <- data.frame(sampleID = "covid19_heidelberg_SER_MS-AA_PAI05_COVp88_261121_QC04_29", project = "COVID-19", cohort = "Heidelberg",
                   projectCohort = "COVID-19 Heidelberg", method = "MS-AA-POS", instrument = "PAI05", plate = "88")
anonymiseMetadata(data = data, sampleID = "sampleID", project = "project", cohort = "cohort", projectCohort = "projectCohort",
                  method = "method", instrument = "instrument", plate = "plate")
}
