% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateAverageMassAccuracy.R
\name{calculateAverageMassAccuracy}
\alias{calculateAverageMassAccuracy}
\title{Calculate Average Mass Accuracy}
\usage{
calculateAverageMassAccuracy(
  data,
  commonColumn,
  spectrumCount,
  firstColumnIndex,
  lastColumnIndex = NULL,
  roundDecimalPlace = NULL
)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{commonColumn}{Column names of data common for each sample: character vector}

\item{spectrumCount}{Spectrum count column name: character}

\item{firstColumnIndex}{Index of the first data column to averaged: double}

\item{lastColumnIndex}{Index of the last data column to averaged: double}

\item{roundDecimalPlace}{Number of decimal places for precision value rounding (Default: NULL): NULL or double}
}
\value{
A data frame grouped by common columns and the average ppm accuracy for each target m/z value
}
\description{
Calculate the average mass accuracy (in ppm) of each target m/z from multiple data columns by grouping using its common columns.
}
