% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateMassAccuracy.R
\name{calculateMassAccuracy}
\alias{calculateMassAccuracy}
\title{Calculate Mass Accuracy}
\usage{
calculateMassAccuracy(data, mzArray, targetMZ, roundDecimal = NULL)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{mzArray}{m/z array column name: character}

\item{targetMZ}{Target m/z values (Must be unique, sorted, and equal length and correspond to the values stored within intensity array): character vector}

\item{roundDecimal}{Number of decimal places for precision value rounding (Default: NULL): NULL or double}

\item{commonColumn}{Column names of data common for each sample: character vector}
}
\value{
A data frame with appended columns representing the calculated mass accuracy (in ppm) for each target m/z (EIC)
}
\description{
Calculate the mass accuracy (in ppm) of each target m/z data detected within the spectra.
}
\details{
The length of the target list containing unique m/z values must be equal to the length of m/z array for each spectral data and reflect the corresponding m/z values stored.
}
