% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSpectrumCount.R
\name{checkSpectrumCount}
\alias{checkSpectrumCount}
\title{Check Spectrum Count}
\usage{
checkSpectrumCount(data, commonColumn, sampleID, spectrumCount, threshold = 20)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{commonColumn}{Column names representing common data to be used for data grouping: character}

\item{sampleID}{Sample ID column name: character}

\item{spectrumCount}{Spectrum count column name: character}

\item{threshold}{Accepted threshold limit \% (Default: 20): numeric}
}
\value{
A list containing the statistical result (sample size n, sum and mean of spectrum count, and the accepted threshold limit), and two data frames (passed and failed data)
}
\description{
Check the total number of spectral data (spectrum count) of each sample against an accepted threshold limit (\%) from the average mean.
}
