% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSpectrumCount.R
\name{checkSpectrumCount}
\alias{checkSpectrumCount}
\title{checkSpectrumCount}
\usage{
checkSpectrumCount(data, commonColumn, sampleID, spectrumCount, threshold = 20)
}
\arguments{
\item{data}{Data frame that contains spectral information of samples: data frame}

\item{commonColumn}{Column names that are common across all samples: character vector}

\item{sampleID}{Name of the column that represents sample ID: character}

\item{spectrumCount}{Name of the column that represents spectrum count: character}

\item{threshold}{Accepted threshold limit \% (Default: 20): double}
}
\value{
List containing the statistical result (sample size, total and mean of spectrum count, and the threshold limit value), and two data frames (passed and failed data)
}
\description{
Check and filter samples against a specific threshold limit set on the average mean of spectral data (spectrum count) present.
}
