% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSpectrumCount.R
\name{checkSpectrumCount}
\alias{checkSpectrumCount}
\title{Check Spectrum Count}
\usage{
checkSpectrumCount(data, commonColumn, sampleID, spectrumCount, threshold = 20)
}
\arguments{
\item{data}{A data frame of the MS spectral data.}

\item{commonColumn}{A character vector representing the names of the common data columns to be used for data grouping.}

\item{sampleID}{A character string representing the name of the sample ID column.}

\item{spectrumCount}{A character string representing the name of the spectrum count column.}

\item{threshold}{A numeric value representing the accepted threshold limit (\%). (Default: \code{20})}
}
\value{
A list with statistical information (sample size, sum, and mean) and two data frames (data that passed and failed the spectrum count check).
}
\description{
Check the total number of MS spectral data (spectrum count) of each data group against an accepted threshold limit (\%) (+/-) from the average mean.
}
