% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUndefinedSampleType.R
\name{checkUndefinedSampleType}
\alias{checkUndefinedSampleType}
\title{Check Undefined Sample Type}
\usage{
checkUndefinedSampleType(data, commonColumn, sampleType, spectrumCount)
}
\arguments{
\item{data}{A data frame containing MS spectral data.}

\item{commonColumn}{A character vector representing the names of the common data columns to be used for data grouping.}

\item{sampleType}{A character string representing the name of the sample type column.}

\item{spectrumCount}{A character string representing the name of the spectrum count column.}
}
\value{
A list with two data frames (data that passed and failed the undefined sample type check).
}
\description{
Check data for undefined sample type.
}
