% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUndefinedSampleType.R
\name{checkUndefinedSampleType}
\alias{checkUndefinedSampleType}
\title{Check Undefined Sample Type}
\usage{
checkUndefinedSampleType(data, commonColumn, sampleType, spectrumCount)
}
\arguments{
\item{data}{A data frame containing spectral data.}

\item{commonColumn}{A character vector representing names of the common data columns to be used for data grouping.}

\item{sampleType}{A character string representing the name of the sample type column.}

\item{spectrumCount}{A character string representing the name of the spectrum count column.}
}
\value{
A list containing two data frames (\code{passedData} containing spectral data of samples with a defined sample type and \code{failedData} containing a summarised listing of samples with an undefined sample type).
}
\description{
Check for samples that have an undefined sample type.
}
