% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUndefinedSampleType.R
\name{checkUndefinedSampleType}
\alias{checkUndefinedSampleType}
\title{checkUndefinedSampleType}
\usage{
checkUndefinedSampleType(data, commonColumn, sampleType, spectrumCount)
}
\arguments{
\item{data}{Data frame that contains spectral information of samples: data frame}

\item{commonColumn}{A vector of column names that are common across all samples: character vector}

\item{sampleType}{The name of the column that represents sample type: character}

\item{spectrumCount}{The name of the column that represents spectrum count: character}
}
\value{
A list containing two data frames (samples that passed and failed the check)
}
\description{
Check and filter for samples that have an undefined sample type.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
