% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkUndefinedSampleType.R
\name{checkUndefinedSampleType}
\alias{checkUndefinedSampleType}
\title{Check Undefined Sample Type}
\usage{
checkUndefinedSampleType(data, commonColumn, sampleType, spectrumCount)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{commonColumn}{Column names of data common for each sample: character or character vector}

\item{sampleType}{Sample type column name: character}

\item{spectrumCount}{Spectrum count column name: character}
}
\value{
A list containing two data frames (passed and failed data)
}
\description{
Check for samples that have an undefined sample type.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
