% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureChromatogramRegion.R
\name{configureChromatogramRegion}
\alias{configureChromatogramRegion}
\title{Configure Chromatogram Region}
\usage{
configureChromatogramRegion(
  massCalibrationStart = 0,
  massCalibrationEnd,
  analyteStart = massCalibrationEnd,
  analyteEnd = washStart,
  washStart,
  washEnd = NULL
)
}
\arguments{
\item{analyteStart}{A numeric value representing the start of the analyte region. (Default: \code{massCalEnd})}

\item{analyteEnd}{A numeric value representing the end of the analyte region. (Default: \code{washStart})}

\item{washStart}{A numeric value representing the start of the wash region.}

\item{washEnd}{A numeric value representing the end of the wash region. (Default: \code{NULL})}

\item{massCalStart}{A numeric value representing the start of the mass calibration region. (Default: \code{0})}

\item{massCalEnd}{A numeric value representing the end of the mass calibration region.}
}
\value{
A list of chromatogram region data of interests.
}
\description{
Configure the start and end time points of the chromatogram regions of interest:
\itemize{
\item Prewash
\item Mass calibration
\item Analyte
\item Wash
}
}
\examples{
# Example 1: Using default values
configureChromatogramRegion(massCalibrationEnd = 0.3, washStart = 5)

# Example 2: Specifying values
configureChromatogramRegion(massCalibrationStart = 0, massCalibrationEnd = 0.3, analyteStart = 2, analyteEnd = 4, washStart = 5, washEnd = 6)
}
