% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureChromatogramRegion.R
\name{configureChromatogramRegion}
\alias{configureChromatogramRegion}
\title{Configure Chromatogram Region}
\usage{
configureChromatogramRegion(
  massCalibrationStart = 0,
  massCalibrationEnd,
  analyteStart = massCalibrationEnd,
  analyteEnd = washStart,
  washStart,
  washEnd = NULL
)
}
\arguments{
\item{analyteStart}{Start of feature detection cycle (Default: End of mass calibration cycle): double}

\item{analyteEnd}{End of feature detection cycle (Default: Start of wash cycle): double}

\item{washStart}{Start of wash cycle: double}

\item{washEnd}{End of wash cycle (Default: NULL): NULL or double}

\item{massCalStart}{Start of mass calibration cycle (Default: 0): double}

\item{massCalEnd}{End of mass calibration cycle: double}
}
\value{
A list containing the start and end time points of the different chromatogram regions of interest (prewash, mass calibration, analyte, and wash)
}
\description{
Configure the start and end time points of the different chromatogram regions of interest:
\itemize{
\item Prewash region
\item Mass calibration region
\item Analyte region
\item Wash region
}
}
\examples{
# Example 1: Using default values for analyteStart, analyteEnd and washEnd
chromatogramRegion <- ticq::configureChromatogramRegion(massCalibrationStart = 0, massCalibrationEnd = 0.3, washStart = 5)
print(chromatogramRegion)

# Example 2: Defining values for analyteStart, analyteEnd and washEnd
chromatogramRegion <- ticq::configureChromatogramRegion(massCalibrationStart = 0, massCalibrationEnd = 0.3,
                                                        analyteStart = 2, analyteEnd = 4,
                                                        washStart = 5, washEnd = 6)
print(chromatogramRegion)
}
