% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureChromatogramRegion.R
\name{configureChromatogramRegion}
\alias{configureChromatogramRegion}
\title{configureChromatogramRegion}
\usage{
configureChromatogramRegion(
  massCalStart = 0,
  massCalEnd,
  analyteStart = massCalEnd,
  analyteEnd = washStart,
  washStart,
  washEnd = NULL
)
}
\arguments{
\item{massCalStart}{Fixed time point when the MS instrument starts its mass calibration cycle (Default: 0): double}

\item{massCalEnd}{Fixed time point that the MS instrument ends its mass calibration cycle: double}

\item{analyteStart}{Fixed time point when the MS instrument starts its cycle to scan for targeted features (Default: End of mass calibration cycle): double}

\item{analyteEnd}{Fixed time point when the MS instrument ends its cycle to scan for targeted features (Default: Start of wash cycle): double}

\item{washStart}{Fixed time point that the MS instrument starts its wash cycle: double}

\item{washEnd}{Fixed time point that the MS instrument ends its wash cycle (Default: NULL): NULL or double}
}
\value{
List containing the time points representing the different chromatogram regions of interest
}
\description{
Configure chromatogram regions of interest based on specific time points.
}
\examples{
# Example 1: Using default values for analyteStart, analyteEnd and washEnd
chromatogramRegion <- ticq::configureChromatogramRegion(massCalStart = 0, massCalEnd = 0.3, 
                                                        washStart = 5)

# Example 2: Define values for analyteStart, analyteEnd and washEnd
chromatogramRegion <- ticq::configureChromatogramRegion(massCalStart = 0, massCalEnd = 0.3,
                                                        analyteStart = 2, analyteEnd = 4
                                                        washStart = 5, washEnd = 6)
}
