% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSpectrum.R
\name{countSpectrum}
\alias{countSpectrum}
\title{Count Spectrum}
\usage{
countSpectrum(data, commonColumn, spectrumCount)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{commonColumn}{Column names representing common data to be used for data grouping: character}

\item{spectrumCount}{Spectrum count column name: character}
}
\value{
A summarised data frame grouped by common columns and their respective total number of spectrum count
}
\description{
Count the total number of spectral data (spectrum count) available for each data group.
}
