% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSpectrum.R
\name{countSpectrum}
\alias{countSpectrum}
\title{countSpectrum}
\usage{
countSpectrum(data, commonColumn, spectrumCount)
}
\arguments{
\item{data}{Data frame that contains spectral information of samples: data frame}

\item{commonColumn}{A vector of column names that are common across all samples: character vector}

\item{spectrumCount}{The name of the column that represents spectrum count: character}
}
\value{
A summarised data frame grouped by common columns and their respective number of spectrum available
}
\description{
Count and summarise the total number of spectral data for each sample present within the data frame.
}
