% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countSpectrum.R
\name{countSpectrum}
\alias{countSpectrum}
\title{Count Spectrum}
\usage{
countSpectrum(data, commonColumn, spectrumCount)
}
\arguments{
\item{data}{A data frame containing MS spectral data.}

\item{commonColumn}{A character vector representing the names of the common data columns to be used for data grouping.}

\item{spectrumCount}{A character string representing the name of the spectrum count column.}
}
\value{
A data frame of MS spectral data summary grouped by their common data columns and its associated spectrum count.
}
\description{
Count the total number of MS spectral data (spectrum count) of each data group.
}
