% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{createDirectoryPath}
\alias{createDirectoryPath}
\title{Create Directory Path}
\usage{
createDirectoryPath(name, value)
}
\arguments{
\item{name}{A character string representing the name of the object being validated.}

\item{value}{A character string representing a directory path.}
}
\value{
Invisible TRUE if the directory exist or is successfully created at the specified path; otherwise, it stops with an error message.
}
\description{
Validate and create a directory based on a character string representing a directory path.
}
