% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHelper.R
\name{createDirectoryPath}
\alias{createDirectoryPath}
\title{Create Directory Path}
\usage{
createDirectoryPath(parameterName, parameterValue)
}
\arguments{
\item{parameterName}{A character string representing the name of the parameter being validated.}

\item{parameterValue}{A character string representing a directory path to be created.}
}
\value{
Invisible TRUE if the directory exists or is created at the specified path; otherwise, it stops with an error message.
}
\description{
A helper function to create a directory using a character string representing a directory path.
It ensures that the directory path specified has a length of 1, is a character type, is not NA or an empty character string, ends with a forward slash (/),
and creates the directory at the specified path.
}
