% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayChromatogram.R
\name{displayChromatogram}
\alias{displayChromatogram}
\title{Display Chromatogram}
\usage{
displayChromatogram(
  data,
  x,
  y,
  colour = NULL,
  title = "Chromatogram",
  subtitle = NULL,
  caption = NULL,
  xLabel = x,
  yLabel = y,
  colourLabel = colour,
  xTickToggle = TRUE,
  facetWrapBy = NULL,
  facetColumn = NULL,
  facetRow = NULL,
  chromatogramRegion = NULL
)
}
\arguments{
\item{data}{A data frame containing plot data.}

\item{x}{A character string representing the name of the data column to be used for the x-axis.}

\item{y}{A character string representing the name of the data column to be used for the y-axis.}

\item{colour}{A character string representing the name of the data column to be used for colour grouping. (Default: \code{NULL})}

\item{title}{A character string representing the plot title. (Default: \code{"Chromatogram"})}

\item{subtitle}{A character string representing the plot subtitle. (Default: \code{NULL})}

\item{caption}{A character string representing the plot caption. (Default: \code{NULL})}

\item{xLabel}{A character string representing the x-axis label. (Default: Value used for \code{x})}

\item{yLabel}{A character string representing the y-axis label. (Default: Value used for \code{y})}

\item{colourLabel}{A character string representing the colour grouping label. (Default: Value used for \code{colour})}

\item{xTickToggle}{A logical value representing the toggle to display ticks on the x-axis. (Default: \code{TRUE}, Options: \code{TRUE} or \code{FALSE})}

\item{facetWrapBy}{A character string representing the name of the data column to be used for facet wrapping. (Default: \code{NULL})}

\item{facetColumn}{A numeric value representing the number of columns to be used for facet wrapping. (Default: \code{NULL})}

\item{facetRow}{A numeric value representing the number of rows to be used for face wrapping. (Default: \code{NULL})}

\item{chromatogramRegion}{A list of lists representing the different chromatogram regions of interest and their respective start and end time points. (Default: \code{NULL}; Options: Use \code{ticq::configureChromatogramRegion()})}
}
\value{
This function does not return any value. It prints the ggplot object displaying the chromatogram.
}
\description{
Display chromatogram data.
}
