% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayChromatogramRegion.R
\name{displayChromatogramRegion}
\alias{displayChromatogramRegion}
\title{Display Chromatogram Region}
\usage{
displayChromatogramRegion(plot, maxX, maxY, regionOfInterest, label = NULL)
}
\arguments{
\item{plot}{A ggplot object of a chromatogram: ggplot}

\item{maxX}{x-axis maximum value: numeric}

\item{maxY}{y-axis maximum value: numeric}

\item{regionOfInterest}{A list representing the start and end time points of a chromatogram region (Options: ticq::configureChromatogram()$region): list}

\item{label}{Chromatogram region label (Default: NULL): NULL or character}
}
\value{
A ggplot object with the added vertical lines representing the chromatogram region start and end time points
}
\description{
Display vertical lines representing the start and end time points of a chromatogram region.
}
