% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayChromatogramRegion.R
\name{displayChromatogramRegion}
\alias{displayChromatogramRegion}
\title{displayChromatogramRegion}
\usage{
displayChromatogramRegion(plot, maxX, maxY, chromatogramRegion, label)
}
\arguments{
\item{plot}{A ggplot that represents a chromatogram: ggplot}

\item{maxX}{The maximum value of the x-axis: double}

\item{maxY}{The maximum value of the y-axis: double}

\item{chromatogramRegion}{The chromatogram region to be displayed: list}

\item{label}{The region label to be displayed: character}
}
\value{
ggplot with the added vertical lines representing the chromatogram region
}
\description{
Add vertical lines to represent the different regions of the chromatogram.
}
\details{
This function is not exported as it is only used by:
ticq::plotChromatogram()
}
