% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayChromatogramRegion.R
\name{displayChromatogramRegion}
\alias{displayChromatogramRegion}
\title{displayChromatogramRegion}
\usage{
displayChromatogramRegion(plot, maxX, maxY, chromatogramRegion, label)
}
\arguments{
\item{plot}{A ggplot object representing a chromatogram: ggplot}

\item{maxX}{x-axis maximum value: double}

\item{maxY}{y-axis maximum value: double}

\item{chromatogramRegion}{Chromatogram region to be displayed: list}

\item{label}{Chromatogram region label: character}
}
\value{
A ggplot object with the added vertical lines displaying the chromatogram region start and end time points
}
\description{
Add vertical lines to display the start and end time points of a chromatogram region.
}
\details{
This function is not exported as it is only used by:
ticq::plotChromatogram()
}
