% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayChromatogramRegion.R
\name{displayChromatogramRegion}
\alias{displayChromatogramRegion}
\title{displayChromatogramRegion}
\usage{
displayChromatogramRegion(plot, maxX, maxY, chromatogramRegion, label)
}
\arguments{
\item{plot}{ggplot object that represents a chromatogram: ggplot}

\item{maxX}{Maximum value of the x-axis: double}

\item{maxY}{Maximum value of the y-axis: double}

\item{chromatogramRegion}{Chromatogram region to be displayed: list}

\item{label}{Chromatogram region label: character}
}
\value{
ggplot object with the added vertical lines displaying the chromatogram region start and end points
}
\description{
Add vertical lines to represent the different regions of the chromatogram.
}
\details{
This function is not exported as it is only used by:
ticq::plotChromatogram()
}
