% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayMultipleChromatogram.R
\name{displayMultipleChromatogram}
\alias{displayMultipleChromatogram}
\title{Display Multiple Chromatogram}
\usage{
displayMultipleChromatogram(
  data,
  x,
  firstColumnIndex,
  lastColumnIndex = NULL,
  colour = NULL,
  title = "Chromatogram",
  subtitle = NULL,
  caption = NULL,
  xLabel = x,
  yLabel = NULL,
  colourLabel = colour,
  xTickToggle = TRUE,
  facetWrapBy = NULL,
  facetColumn = NULL,
  facetRow = NULL,
  plotColumn = 1,
  plotRow = 1,
  chromatogramRegion = NULL
)
}
\arguments{
\item{data}{A data frame of the chromatogram data to be used for plotting.}

\item{x}{A character string representing the name of the data column to be used for the x-axis.}

\item{firstColumnIndex}{A numeric value representing the index of the first data column to be used for y-axis.}

\item{lastColumnIndex}{A numeric value representing the index of the last data column to be used for y-axis. (Default: \code{NULL})}

\item{colour}{A character string representing the name of the data column to be used for colour grouping. (Default: \code{NULL})}

\item{title}{A character string representing the plot title. (Default: \code{"Chromatogram"})}

\item{subtitle}{A character string representing the plot subtitle. (Default: \code{NULL})}

\item{caption}{A character string representing the plot caption. (Default: \code{NULL})}

\item{xLabel}{A character string representing the x-axis label. (Default: \code{x})}

\item{yLabel}{A character string representing the y-axis label. (Default: \code{y})}

\item{colourLabel}{A character string representing the colour grouping label. (Default: \code{colour})}

\item{xTickToggle}{A logical value representing the toggle to display the ticks on the x-axis. (Default: \code{TRUE}, Options: \code{TRUE} or \code{FALSE})}

\item{facetWrapBy}{A character string representing the name of the data column to be used for facet wrapping. (Default: \code{NULL})}

\item{facetColumn}{A numeric value representing the number of columns to be used for facet wrapping. (Default: \code{NULL})}

\item{facetRow}{A numeric value representing the number of rows to be used for face wrapping. (Default: \code{NULL})}

\item{plotColumn}{A numeric value representing the number of columns to be used for plotting (Default: \code{1})}

\item{plotRow}{A numeric value representing the number of rows to be used for plotting (Default: \code{1})}

\item{chromatogramRegion}{A list representing the chromatogram region data of interests. (Default: \code{NULL}; Options: \code{configureChromatogramRegion()})}
}
\value{
This function does not return any value. It prints the ggplot objects displaying the multiple chromatogram.
}
\description{
Display multiple chromatogram data.
}
\details{
Chromatogram region can be configured using ticq::configureChromatogramRegion().
}
