% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayPCA.R
\name{displayPCA}
\alias{displayPCA}
\title{Display Principal Component Analysis (PCA)}
\usage{
displayPCA(
  data,
  firstColumnIndex,
  lastColumnIndex = NULL,
  scale = TRUE,
  confidence = 95,
  distribution = NULL,
  colour = NULL,
  shape = NULL,
  subtitle = NULL,
  colourLabel = colour,
  shapeLabel = shape,
  screePlotToggle = TRUE,
  scoresPlotToggle = TRUE,
  loadingsPlotToggle = TRUE,
  biPlotToggle = TRUE
)
}
\arguments{
\item{data}{A data frame containing data to perform PCA on (Must be equal sized and should not contain NA values): data frame}

\item{firstColumnIndex}{Index of the first data column to perform PCA on: numeric}

\item{lastColumnIndex}{Index of the last data column data to perform PCA on (Default: NULL): NULL or numeric}

\item{scale}{PCA scaling (Default: TRUE, Options: TRUE or FALSE): logical}

\item{confidence}{Data confidence \% (Default: 95, Options: 95 or 99): numeric}

\item{distribution}{Data distribution to be drawn on scores plot and biplot (Default: NULL, Options: "normal" or "t"): NULL or character}

\item{colour}{Column name representing data to be used for colour grouping (Default: NULL): NULL or character}

\item{shape}{Column name representing data to be used for shape grouping (Default: NULL): NULL or character}

\item{subtitle}{Plot subtitle (Default: NULL): NULL or character}

\item{colourLabel}{Colour grouping label (Default: Value used for colour): NULL or character}

\item{shapeLabel}{Shape grouping label (Default: Value used for shape): NULL or character}

\item{screePlotToggle}{Toggle to display scree plot (Default: TRUE, Options: TRUE or FALSE): logical}

\item{scoresPlotToggle}{Toggle to display scores plot (Default: TRUE, Options: TRUE or FALSE): logical}

\item{loadingsPlotToggle}{Toggle to display loadings plot (Default: TRUE, Options: TRUE or FALSE): logical}

\item{biPlotToggle}{Toggle to display biplot (Default: TRUE, Options: TRUE or FALSE): logical}
}
\description{
Perform Principle Component Analysis (PCA) and display analysis outcomes (scree plot, scores plot, loadings plot and/or biplot).
}
