% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayViolinBox.R
\name{displayViolinBox}
\alias{displayViolinBox}
\title{Display Violin/Box Plot}
\usage{
displayViolinBox(
  data,
  x,
  y,
  colour = NULL,
  shape = NULL,
  outlierShape = NA,
  title = "Violin/Box Plot",
  subtitle = NULL,
  caption = NULL,
  xLabel = x,
  yLabel = y,
  colourLabel = colour,
  shapeLabel = shape,
  boundary = NULL,
  boundaryValue = 0,
  referenceData = NULL,
  violinPlotToggle = TRUE,
  tailTrimToggle = FALSE,
  boxPlotToggle = TRUE,
  boxWidth = 1,
  facetWrapBy = NULL,
  facetColumn = NULL,
  facetRow = NULL
)
}
\arguments{
\item{data}{A data frame containing plot data: data frame}

\item{x}{Column name representing the continuous data series to be used for x-axis: character}

\item{y}{Column name representing the continuous data series to be used for y-axis: character}

\item{colour}{Column name representing data to be used for colour grouping (Default: NULL): NULL or character}

\item{shape}{Column name representing data to be used for shape grouping (Default: NULL): NULL or character}

\item{outlierShape}{Shape style to be used to identify outliers (Default: NA): NA or double}

\item{title}{Plot title (Default: "Violin/Box Plot"): NULL or character}

\item{subtitle}{Plot subtitle (Default: NULL): NULL or character}

\item{caption}{Plot caption (Default: NULL): NULL or character}

\item{xLabel}{x-axis label (Default: Value used for x): NULL or character}

\item{yLabel}{y-axis label (Default: Value used for y): NULL or character}

\item{colourLabel}{Colour grouping label (Default: Value used for colour): NULL or character}

\item{shapeLabel}{Shape grouping label (Default: Value used for shape)}

\item{boundary}{Boundary analysis to perform (Default: NULL, Options: "value", "percentage", "sd"): NULL or character.}

\item{boundaryValue}{Value used for "value" and "percentage" boundary analysis (Default: 0): double}

\item{referenceData}{A data frame with the same column names and data to calculate "sd" boundaries (Default: NULL): NULL or data frame}

\item{violinPlotToggle}{Toggle to display violin plot (Default: TRUE, Options: TRUE or FALSE): boolean}

\item{tailTrimToggle}{Toggle to trim the tails of the violin plot. Applicable only if violinPlotToggle is set to TRUE (Default: FALSE, Options: TRUE or FALSE): boolean}

\item{boxPlotToggle}{Toggle to display box plot (Default: TRUE, Options: TRUE or FALSE): boolean}

\item{boxWidth}{Box plot width. Applicable only if boxPlotToggle is set to TRUE (Default: 1): double}

\item{facetWrapBy}{Column name representing data to be used for facet wrapping (Default: NULL): NULL or character}

\item{facetColumn}{Number of columns to use for facet wrapping (Default: NULL): NULL or double}

\item{facetRow}{Number of rows to use for facet wrapping (Default: NULL): NULL or double}
}
\value{
A data frame containing the result of the boundary analysis performed
}
\description{
Display data distribution using violin and/or box plot (and perform boundary analysis if defined).
}
\details{
Boundary analysis that can be performed:
\itemize{
\item value: Evaluates data by +/- value from 0
\item percentage: Evaluates data +/- value from 100
\item sd: Evaluates data using 2 standard deviations from average mean of the data or using defined mean and standard deviation values
}

To define mean and standard deviation values to be used for "sd" boundary analysis, pass reference data that mirrors the same columns as data.
}
