% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCohort.R
\name{extractCohort}
\alias{extractCohort}
\title{extractCohort}
\usage{
extractCohort(input)
}
\arguments{
\item{input}{Input to extract cohort name from: character}
}
\value{
Cohort name
}
\description{
Extract cohort name.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
\examples{
cohort <- ticq::extractCohort(input = "covid19_heidelberg_SER_MS-AA_PAI05_COVp88_261121_QC04_29.json")
}
