% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCohort.R
\name{extractCohort}
\alias{extractCohort}
\title{Extract Cohort}
\usage{
extractCohort(input)
}
\arguments{
\item{input}{A character vector representing ANPC-specific data file names or paths.}
}
\value{
A character vector of the associated cohort names.
}
\description{
Extract cohort name.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
