% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEIC.R
\name{extractEIC}
\alias{extractEIC}
\title{Extract Extracted Ion Current (EIC)}
\usage{
extractEIC(data, intensityArray, targetMZ)
}
\arguments{
\item{data}{A data frame of the MS spectral data.}

\item{intensityArray}{A character string representing the name of the intensity array column.}

\item{targetMZ}{A character vector representing the targeted m/z values in the spectra array.}
}
\value{
A data frame of the MS spectral data and the intensity of each target m/z value.
}
\description{
Extract the intensity of each target m/z value in the spectra array.
}
