% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEIC.R
\name{extractEIC}
\alias{extractEIC}
\title{extractEIC}
\usage{
extractEIC(data, intensityArray, targetList)
}
\arguments{
\item{data}{Data frame that contains spectral information of samples: data frame}

\item{intensityArray}{Name of the column that represents intensity array: character}

\item{targetList}{Unique target m/z value that reflect the intensity value in the array: character vector}
}
\value{
Data frame with the extracted intensity of each target m/z (EIC) in its own individual column appended
}
\description{
Extract the intensity of each target m/z (Extracted Ion Current, EIC) stored within the spectra.
}
