% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEIC.R
\name{extractEIC}
\alias{extractEIC}
\title{Extract Extracted Ion Current (EIC)}
\usage{
extractEIC(data, intensityArray, targetList)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{intensityArray}{Intensity array column name: character}

\item{targetList}{A list of unique target m/z mirroring the values stored within the intensity array (Must be equal length): character vector}
}
\value{
A data frame with appended columns representing the extracted intensity of each target m/z (EIC)
}
\description{
Extract the intensity of each target m/z (EIC) detected within the spectra.
}
\details{
The length of the target list of unique m/z values MUST BE EQUAL to the length of spectra array (m/z and intensity)
}
