% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExfilMS.R
\name{extractExfilMS}
\alias{extractExfilMS}
\title{Extract ExfilMS}
\usage{
extractExfilMS(inputPath)
}
\arguments{
\item{inputPath}{Input path to JSON files generated by ExfilMS: character}
}
\value{
A list containing two data frames (passed and failed data)
}
\description{
Extract spectral data from JSON files generated using ExfilMS (https://github.com/vmalnathnambiar/exfilms). Also includes retrievable metadata from sample name or file path.
}
\details{
Metadata extraction only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
