% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExfilMS.R
\name{extractExfilMS}
\alias{extractExfilMS}
\title{Extract ExfilMS}
\usage{
extractExfilMS(inputPath)
}
\arguments{
\item{inputPath}{A character string representing a directory or file path containing ExfilMS extracted JSON files.}
}
\value{
A list with two data frames (data that passed and failed the extraction process).
}
\description{
Extract MS spectral data along with its associated metadata from \href{https://github.com/vmalnathnambiar/exfilms#readme}{ExfilMS} extracted JSON files.
}
\details{
Metadata extraction only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
