% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractExfilMS.R
\name{extractExfilMS}
\alias{extractExfilMS}
\title{Extract ExfilMS}
\usage{
extractExfilMS(inputPath)
}
\arguments{
\item{inputPath}{A character string representing an existing directory or file path.}
}
\value{
A list containing two data frames (\code{passedData} containing the extracted spectral data and \code{failedData} with file names that failed the extraction process).
}
\description{
Extract spectral data and retrieve metadata from JSON files generated using \href{https://github.com/vmalnathnambiar/exfilms#readme}{ExfilMS}.
}
\details{
Metadata extraction only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
