% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInstrument.R
\name{extractInstrument}
\alias{extractInstrument}
\title{Extract Instrument}
\usage{
extractInstrument(input)
}
\arguments{
\item{input}{ANPC sample file name or path: character}
}
\value{
Instrument name
}
\description{
Extract instrument name.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
