% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInstrument.R
\name{extractInstrument}
\alias{extractInstrument}
\title{extractInstrument}
\usage{
extractInstrument(input)
}
\arguments{
\item{input}{Input to extract instrument name: character}
}
\value{
Instrument name
}
\description{
Extract instrument name.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
\examples{
instrument <- ticq::extractInstrument(input = "covid19_heidelberg_SER_MS-AA_PAI05_COVp88_261121_QC04_29.json")
}
