% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMatrix.R
\name{extractMatrix}
\alias{extractMatrix}
\title{Extract Matrix}
\usage{
extractMatrix(input)
}
\arguments{
\item{input}{A character vector representing ANPC-specific data file names or paths.}
}
\value{
A character vector of the associated matrix types.
}
\description{
Extract matrix type.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
