% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMatrix.R
\name{extractMatrix}
\alias{extractMatrix}
\title{extractMatrix}
\usage{
extractMatrix(input)
}
\arguments{
\item{input}{Input to extract matrix type from: character}
}
\value{
Matrix type
}
\description{
Extract matrix type.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
\examples{
matrix <- ticq::extractMatrix(input = "covid19_heidelberg_SER_MS-AA_PAI05_COVp88_261121_QC04_29.json")
}
