% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{extractMetadata}
\alias{extractMetadata}
\title{Extract Metadata}
\usage{
extractMetadata(input)
}
\arguments{
\item{input}{ANPC sample file name or path: character or character vector}
}
\value{
List containing the associated metadata associated of the input
}
\description{
Extract sample metadata including project name, cohort name, matrix type, sample type, acquisition method, instrument name, and plate number.
}
\details{
Applicable only to data acquired at the Australian National Phenome Centre (ANPC).
}
\examples{
metadata <- ticq::extractMetadata(input = "covid19_heidelberg_SER_MS-AA_PAI05_COVp88_261121_QC04_29.json")
print(metadata)
}
