% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetadata.R
\name{extractMetadata}
\alias{extractMetadata}
\title{extractMetadata}
\usage{
extractMetadata(input)
}
\arguments{
\item{input}{ANPC sample file name or path: character}
}
\value{
List containing the metadata associated to the data
}
\description{
Extract metadata. This includes project name, cohort name, matrix type, sample type, acquisition method, instrument name, and plate number.
}
\details{
Applicable only to data acquired at the Australian National Phenome Centre (ANPC).
}
\examples{
metadata <- ticq::extractMetadata(input = "covid19_heidelberg_SER_MS-AA_PAI05_COVp88_261121_QC04_29.json")
print(metadata)
}
