% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMethod.R
\name{extractMethod}
\alias{extractMethod}
\title{extractMethod}
\usage{
extractMethod(input)
}
\arguments{
\item{input}{Input to extract method name: character}
}
\value{
Method name
}
\description{
Extract method name.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
\examples{
method <- ticq::extractMethod(input = "covid19_heidelberg_SER_MS-AA_PAI05_COVp88_261121_QC04_29.json")
print(method)
}
