% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMethod.R
\name{extractMethod}
\alias{extractMethod}
\title{Extract Method}
\usage{
extractMethod(input)
}
\arguments{
\item{input}{A character vector representing ANPC sample file names or paths.}
}
\value{
A character vector representing method names.
}
\description{
Extract method name.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
