% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPlate.R
\name{extractPlate}
\alias{extractPlate}
\title{Extract Plate}
\usage{
extractPlate(input)
}
\arguments{
\item{input}{A character vector representing ANPC-specific data file names or paths.}
}
\value{
A character vector of the associated plate numbers.
}
\description{
Extract plate number.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
