% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPlate.R
\name{extractPlate}
\alias{extractPlate}
\title{extractPlate}
\usage{
extractPlate(input)
}
\arguments{
\item{input}{Input to extract plate number: character}
}
\value{
Plate number
}
\description{
Extract plate number.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
\examples{
plate <- ticq::extractPlate(input = "covid19_heidelberg_SER_MS-AA_PAI05_COVp88_261121_QC04_29.json")
}
