% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractProject.R
\name{extractProject}
\alias{extractProject}
\title{extractProject}
\usage{
extractProject(input)
}
\arguments{
\item{input}{Input to extract project name: character}
}
\value{
Project name
}
\description{
Extract project name.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
\examples{
project <- ticq::extractProject(input = "covid19_heidelberg_SER_MS-AA_PAI05_COVp88_261121_QC04_29.json")
}
