% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSampleType.R
\name{extractSampleType}
\alias{extractSampleType}
\title{Extract Sample Type}
\usage{
extractSampleType(input, matrix = NA_character_)
}
\arguments{
\item{input}{A character vector representing ANPC sample file names or paths.}

\item{matrix}{A character vector representing matrix types. (Default: \code{NA_character_})}
}
\value{
A character vector representing sample types.
}
\description{
Extract sample type.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
