% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSampleType.R
\name{extractSampleType}
\alias{extractSampleType}
\title{Extract Sample Type}
\usage{
extractSampleType(input, matrix = NA_character_)
}
\arguments{
\item{input}{ANPC sample file name or path: character}

\item{matrix}{Associated matrix type (Default: NA_character_): NA_character_ or character}
}
\value{
Sample type
}
\description{
Extract sample type.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
