% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSampleType.R
\name{extractSampleType}
\alias{extractSampleType}
\title{extractSampleType}
\usage{
extractSampleType(input, matrix)
}
\arguments{
\item{input}{Input to extract sample type: character}

\item{matrix}{Associated sample matrix type: character}
}
\value{
Sample type
}
\description{
Extract sample type.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
\examples{
sampleType <- ticq::extractSampleType(input = "covid19_heidelberg_SER_MS-AA_PAI05_COVp88_261121_QC04_29.json",
                                      matrix = "Serum")
print(sampleType)
}
