% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSampleType.R
\name{extractSampleType}
\alias{extractSampleType}
\title{Extract Sample Type}
\usage{
extractSampleType(input, matrix = NA_character_)
}
\arguments{
\item{input}{ANPC sample file name or path: character or character vector}

\item{matrix}{Associated matrix type (Default: NA_character_): NA_character_ or character or character vector}
}
\value{
Sample type
}
\description{
Extract sample type.
}
\details{
Only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
\examples{
sampleType <- ticq::extractSampleType(input = "covid19_heidelberg_SER_MS-AA_PAI05_COVp88_261121_QC04_29.json", matrix = "Serum")
print(sampleType)
}
