% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSpectrum.R
\name{extractSpectrum}
\alias{extractSpectrum}
\title{extractSpectrum}
\usage{
extractSpectrum(input, inputFiles)
}
\arguments{
\item{input}{Input path (can either a directory or a single file path) containing input files for extraction: character}

\item{inputFiles}{Input files to be extracted: character vector}
}
\value{
List containing two data frames (passed and failed data)
}
\description{
Extract spectral data from JSON files generated by ExfilMS (https://github.com/vmalnathnambiar/exfilms) along with the metadata (based on input path).
}
\details{
Metadata extraction only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
