% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSpectrum.R
\name{extractSpectrum}
\alias{extractSpectrum}
\title{extractSpectrum}
\usage{
extractSpectrum(input, inputFiles)
}
\arguments{
\item{input}{Input path (can be a directory or a single file) containing input files for extraction: character}

\item{inputFiles}{A vector of input files for extraction: character vector}
}
\value{
A list containing two data frames (samples that passed and failed the extraction process)
}
\description{
Extract spectral data along with metadata from JSON files generated by ExfilMS (https://github.com/vmalnathnambiar/exfilms).
}
\details{
Metadata extraction only applicable to data acquired at the Australian National Phenome Centre (ANPC).
}
\examples{
input <- "/path/to/input/directory/"
result <- ticq::extractSpectrum(input = input, 
                                inputFiles = ticq::retrieveFileList(input = input, fileExtension = "JSON"))
}
