% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTargetFile.R
\name{extractTargetFile}
\alias{extractTargetFile}
\title{Extract Target File}
\usage{
extractTargetFile(
  targetFilePath = NULL,
  anpcMethodLibrary = NULL,
  roundDecimalPlace = NULL
)
}
\arguments{
\item{targetFilePath}{Published to web URL or local path of the TSV formatted target file (Default: NULL, Options: Published to web URL or local file path): NULL or character}

\item{anpcMethodLibrary}{ANPC method library (Default: NULL, Options: "MS-AA-POS", "MS-HIL-POS", "MS-HIL-NEG", "MS-RP-POS" or "MS-RP-NEG"): NULL or character}

\item{roundDecimalPlace}{Number of decimal places for precision value rounding (Default: NULL): NULL or numeric}
}
\value{
A data frame representing the target file
}
\description{
Extract target m/z data from a TSV formatted target file (published to web URL or local file path) or an in-house method library used at the Australian National Phenome Centre (ANPC).
}
\details{
If both target file path and ANPC method library is specified, the target file path will be used as default extraction. If no data is found, ANPC method library is then used.

When specifying a target file path, please ensure the TSV document follows a pre-determined layout used by ExfilMS (https://github.com/vmalnathnambiar/exfilms/blob/main/docs/how-to-create-a-target-file.md).
}
\examples{
# Example 1: Specifying a target file with no rounding of decimal places for precision values
targetFilePath <- "https://docs.google.com/spreadsheets/d/e/2PACX-1vSeo31hlruA3QuwoESz5IDJ9Nu6ndSAgLTRn3uc45rOPO4BlksfHzh9xtNB22Oes9JOxhEbI4NK-zxl/pub?gid=0&single=true&output=tsv"
targetFile <- ticq::extractTargetFile(targetFilePath = targetFilePath, anpcMethodLibrary = NULL, roundDecimalPlace = NULL)
print(targetFile)

# Example 2: Specifying an ANPC method library with rounding of precision values up to 4 decimal places
targetFile <- ticq::extractTargetFile(targetFilePath = NULL, anpcMethodLibrary = "MS-AA-POS", roundDecimalPlace = 4)
print(targetFile)
}
