% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTargetMZ.R
\name{extractTargetMZ}
\alias{extractTargetMZ}
\title{extractTargetMZ}
\usage{
extractTargetMZ(
  targetFile = NULL,
  anpcMethodLibrary = NULL,
  roundDecimal = NULL
)
}
\arguments{
\item{targetFile}{TSV document that contains target m/z data (Default: NULL, Options: Published to web URL or local file path): NULL or character}

\item{anpcMethodLibrary}{In-house method library used at the ANPC (Default: NULL, Options: "MS-AA-POS", "MS-HIL-POS", "MS-HIL-NEG", "MS-RP-POS" or "MS-RP-NEG"): NULL or character}

\item{roundDecimal}{Number of decimal places to round target m/z precision values (Default: NULL): NULL or double}
}
\value{
Data frame containing the target m/z data
}
\description{
Extract target m/z data from a target file (published to web URL or local TSV document) or an in-house method library used at the Australian National Phenome Centre (ANPC).
}
\examples{
# Example 1: Specify a target file and perform no rounding of m/z precision values
targetMZ <- ticq::extractTargetMZ(targetFile = "https://docs.google.com/spreadsheets/d/e/2PACX-1vSeo31hlruA3QuwoESz5IDJ9Nu6ndSAgLTRn3uc45rOPO4BlksfHzh9xtNB22Oes9JOxhEbI4NK-zxl/pub?gid=0&single=true&output=tsv",
                                  anpcMethodLibrary = NULL,
                                  roundDecimal = NULL)
print(targetMZ)

# Example 2: Specify an in-house method library and round the m/z precision values to 4 decimal places
targetMZ <- ticq::extractTargetMZ(targetFile = NULL,
                                  anpcMethodLibrary = "MS-AA-POS",
                                  roundDecimal = 4)
print(targetMZ)
}
