% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractTargetMZ.R
\name{extractTargetMZ}
\alias{extractTargetMZ}
\title{Extract Target m/z}
\usage{
extractTargetMZ(
  targetFile = NULL,
  anpcMethodLibrary = NULL,
  roundDecimal = NULL
)
}
\arguments{
\item{targetFile}{Published to web or local TSV document containing target m/z data used by ExfilMS for spectra filtering (Default: NULL, Options: Published to web URL or local file path): NULL or character}

\item{anpcMethodLibrary}{ANPC in-house method library (Default: NULL, Options: "MS-AA-POS", "MS-HIL-POS", "MS-HIL-NEG", "MS-RP-POS" or "MS-RP-NEG"): NULL or character}

\item{roundDecimal}{Number of decimal places for precision value rounding (Default: NULL): NULL or double}
}
\value{
A data frame containing target m/z data
}
\description{
Extract target m/z data from a target file (published to web URL or local TSV document) or an in-house method library used at the Australian National Phenome Centre (ANPC).
}
\details{
To assess targeted MS spectral data acquired using ExfilMS post-targeted spectra filtering, ensure the same target file or in-house method library is defined in TICQ.

If both target file and in-house method library is defined, only target file will be checked for. The target file should follow a pre-determined layout used by ExfilMS (https://github.com/vmalnathnambiar/exfilms/blob/main/docs/how-to-create-a-target-file.md).
}
\examples{
# Example 1: Defining a target file with no rounding of precision values
targetMZ <- ticq::extractTargetMZ(targetFile = "https://docs.google.com/spreadsheets/d/e/2PACX-1vSeo31hlruA3QuwoESz5IDJ9Nu6ndSAgLTRn3uc45rOPO4BlksfHzh9xtNB22Oes9JOxhEbI4NK-zxl/pub?gid=0&single=true&output=tsv",
                                  anpcMethodLibrary = NULL,
                                  roundDecimal = NULL)
print(targetMZ)

# Example 2: Defining an in-house method library with rounding of precision values to 4 decimal places
targetMZ <- ticq::extractTargetMZ(targetFile = NULL,
                                  anpcMethodLibrary = "MS-AA-POS",
                                  roundDecimal = 4)
print(targetMZ)
}
