% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterChromatogramRegion.R
\name{filterChromatogramRegion}
\alias{filterChromatogramRegion}
\title{filterChromatogramRegion}
\usage{
filterChromatogramRegion(data, regionOfInterest = NULL, filterBy)
}
\arguments{
\item{data}{Data frame that contains spectral information of samples: data frame}

\item{regionOfInterest}{A list that represent the start and end point of a regions on the chromatogram (Default: NULL): NULL or list}

\item{filterBy}{The name of the column that represents the retention time to be filtered: character}
}
\value{
A data frame containing spliced chromatogram region of interest
}
\description{
Filter for a chromatogram region of interest.
}
