% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterChromatogramRegion.R
\name{filterChromatogramRegion}
\alias{filterChromatogramRegion}
\title{filterChromatogramRegion}
\usage{
filterChromatogramRegion(data, regionOfInterest = NULL, retentionTime)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{regionOfInterest}{A list representing the start and end time points of a chromatogram region (Default: NULL): NULL or list}

\item{retentionTime}{Retention time column name: character}
}
\value{
A data frame containing spectral data that falls within the retention time of the chromatogram region of interest
}
\description{
Filter for retention time of a chromatogram region of interest.
}
