% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterChromatogramRegion.R
\name{filterChromatogramRegion}
\alias{filterChromatogramRegion}
\title{Filter Chromatogram Region}
\usage{
filterChromatogramRegion(data, regionOfInterest = NULL, retentionTime)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{regionOfInterest}{A list representing the start and end time points of a chromatogram region (Default: NULL, Options: Options: ticq::configureChromatogram()$region): NULL or list}

\item{retentionTime}{Retention time column name: character}
}
\value{
A data frame containing spectral data that falls between the start and end time point of a specific chromatogram region of interest
}
\description{
Filter spectral data that falls between the respective start and end time points of a chromatogram region of interest.
}
