% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateStat.R
\name{generateStat}
\alias{generateStat}
\title{Generate Statistics}
\usage{
generateStat(data)
}
\arguments{
\item{data}{A continuous data series of numerical values (Must have a minimum of 2 data points and do not contain NA values): numerical vector}
}
\value{
A list of all generated statistics
}
\description{
Generate statistical information of a dataset including:
\itemize{
\item Sample size (n)
\item Minimum (min)
\item Mean (mean)
\item Maximum (max)
\item 1st Quartile (Q1)
\item Median
\item 3rd Quartile (Q3)
\item Inter-quantile range (IQR)
\item IQR inner fence (Lower and upper)
\item IQR outer fence (Lower and upper)
\item Standard deviation (sd)
\item Standard error (se)
\item Margin error (me) - 95\% and 99\%
\item Confidence interval (CI) - 95\% and 99\% (Lower and upper)
\item Confidence quantile range (CQ) - 95\% and 99\% (Lower and upper)
}
}
\details{
Data must have a minimum of 2 data points and must not contain any NA values.
}
\examples{
stats <- ticq::generateStat(data = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
print(stats)
}
