% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackage.R
\name{loadPackage}
\alias{loadPackage}
\title{loadPackage}
\usage{
loadPackage(packageList)
}
\arguments{
\item{packageList}{Package names to be loaded: character vector}
}
\description{
Identify (and install missing) and load packages to use.
}
\examples{
# Example 1: Loading a single package
ticq::loadPackage(packageList = c("DT"))

# Example 2: Loading multiple packages
ticq::loadPackage(packageList = c("DT", "dplyr"))
}
