% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackage.R
\name{loadPackage}
\alias{loadPackage}
\title{loadPackage}
\usage{
loadPackage(packageList)
}
\arguments{
\item{packageList}{Package names: character vector}
}
\description{
Load packages for usage. Automatically identifies and installs missing packages.
}
\examples{
# Example 1: Load a single package
ticq::loadPackage(packageList = c("DT"))

# Example 2: Load multiple packages
ticq::loadPackage(packageList = c("DT", "dplyr"))
}
