% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPackage.R
\name{loadPackage}
\alias{loadPackage}
\title{Load Package}
\usage{
loadPackage(packageName)
}
\arguments{
\item{packageName}{Package name: character or character vector}
}
\description{
Identify, install and load packages to use.
}
\examples{
# Example 1: Loading a single package
ticq::loadPackage(packageName = "DT")

# Example 2: Loading multiple packages
ticq::loadPackage(packageName = c("DT", "dplyr"))
}
