% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChromatogram.R
\name{plotChromatogram}
\alias{plotChromatogram}
\title{plotChromatogram}
\usage{
plotChromatogram(
  data,
  x,
  y,
  colour = NULL,
  title = "Chromatogram",
  subtitle = NULL,
  caption = NULL,
  xLabel = x,
  yLabel = y,
  colourLabel = colour,
  chromatogramRegion = NULL
)
}
\arguments{
\item{data}{A data frame containing plot data: data frame}

\item{x}{Column name representing the continuous data series to be used for x-axis: character}

\item{y}{Column name representing the continuous data series to be used for y-axis: character}

\item{colour}{Column name representing data to be used for colour grouping (Default: NULL): NULL or character}

\item{title}{Plot title (Default: "Chromatogram"): NULL or character}

\item{subtitle}{Plot subtitle (Default: NULL): NULL or character}

\item{caption}{Plot caption (Default: NULL): NULL or character}

\item{xLabel}{x-axis label (Default: Column name used for x): NULL or character}

\item{yLabel}{y-axis label (Default:Column name used for y): NULL or character}

\item{colourLabel}{Colour grouping label (Default: Value used for colour): NULL or character}

\item{chromatogramRegion}{A list that represents the different chromatogram regions of interest to be plotted (Default: NULL): NULL or list}
}
\description{
Plot chromatogram data.
}
