% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChromatogram.R
\name{plotChromatogram}
\alias{plotChromatogram}
\title{plotChromatogram}
\usage{
plotChromatogram(
  data,
  x,
  y,
  colour = NULL,
  title = "Chromatogram",
  subtitle = NULL,
  caption = NULL,
  labelX = x,
  labelY = y,
  legend = colour,
  chromatogramRegion = NULL
)
}
\arguments{
\item{data}{Data frame that contains plot data: data frame}

\item{x}{Name of the column that represents the continuous data series for the x-axis: character}

\item{y}{Name of the column that represents the continuous data series for the y-axis: character}

\item{colour}{Name of the column used for colour grouping (Default: NULL): NULL or character}

\item{title}{Plot title (Default: "Chromatogram"): character}

\item{subtitle}{Plot subtitle (Default: NULL): NULL or character}

\item{caption}{Plot caption (Default: NULL): NULL or character}

\item{labelX}{x-axis label (Default: Value used for x): NULL or character}

\item{labelY}{y-axis label (Default:Value used for y): NULL or character}

\item{legend}{Legend label for colour grouping (Default: Value used for colour): NULL or character}

\item{chromatogramRegion}{List that represents the different chromatogram regions of interest (Default: NULL): NULL or list}
}
\description{
Plot chromatogram data.
}
