% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultipleChromatogram.R
\name{plotMultipleChromatogram}
\alias{plotMultipleChromatogram}
\title{plotMultipleChromatogram}
\usage{
plotMultipleChromatogram(
  data,
  x,
  startIDX,
  endIDX = NULL,
  colour = NULL,
  caption = NULL,
  labelX = x,
  labelY = NULL,
  colourLegend = colour,
  nPlotCol = 1,
  nPlotRow = 1
)
}
\arguments{
\item{data}{Data frame that contains plot data: data frame}

\item{x}{Name of the column that represents the continuous data series for the x-axis: character}

\item{startIDX}{Index of the first data column that contains continuous data series for the y-axis: double}

\item{endIDX}{Index of the last data column data that contains continuous data series for the y-axis (Default: NULL): NULL or double}

\item{colour}{Name of the column used for colour grouping (Default: NULL): NULL or character}

\item{caption}{Plot caption. This will append a caption value to the column name used for the y-axis (Default: Column name used for y-axis): NULL or character}

\item{labelX}{x-axis label (Default: the name of the column used for the x-axis): NULL or character}

\item{labelY}{y-axis label (Default: NULL): NULL or character}

\item{colourLegend}{Legend label for colour grouping (Default: Value used for colour): NULL or character}

\item{nPlotCol}{Number of plots in a single columns (Default: 1): double}

\item{nPlotRow}{Number of plot in a single row (Default: 1): double}
}
\value{
Print multiple chromatogram plot
}
\description{
Plot multiple chromatogram data in specified number of rows and columns based on the same x-axis but different y-axis.
}
