% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plotPCA}
\alias{plotPCA}
\title{plotPCA}
\usage{
plotPCA(
  data,
  startIDX,
  endIDX = NULL,
  scale = TRUE,
  confidence = 95,
  distribution = NULL,
  colour = NULL,
  subtitle = NULL,
  legend = NULL,
  screePlt = FALSE,
  scoresPlt = FALSE,
  loadingsPlt = FALSE,
  biPlt = FALSE
)
}
\arguments{
\item{data}{An equal size data frame that contains columns representing data (variance) to perform PCA on. Should not contain NA values: data frame}

\item{startIDX}{The index of the first column data to perform PCA on: double}

\item{endIDX}{The index of the last column data to perform PCA on (Default: NULL): NULL or double}

\item{scale}{Scaling status for the PCA (Default: TRUE, Options: TRUE or FALSE): boolean}

\item{confidence}{Confidence value of the data (Default: 95): double}

\item{distribution}{Add distribution boundaries to the PCA scores plot (Default: NULL, Options: "normal" or "t"): NULL or character}

\item{colour}{The name of the column that will be used to colour categorise the data (Default: NULL): NULL or character}

\item{subtitle}{The subtitle for the plot (Default: NULL): NULL or character}

\item{legend}{Legend label that represents the different colour categories (Default: the name of the column used to colour categorise the data): NULL or character}
}
\description{
Plot data generated from Principle Component Analysis (PCA): Scree plot, scores plot, loadings plot and biplot..
}
