% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Plot Time Series}
\usage{
plotTimeSeries(
  data,
  x,
  y,
  colour = NULL,
  shape = NULL,
  title = "Time Series",
  subtitle = NULL,
  caption = NULL,
  xLabel = x,
  yLabel = y,
  colourLabel = colour,
  shapeLabel = shape,
  boundary = NULL,
  boundaryValue = 0,
  referenceData = NULL,
  xTickToggle = TRUE
)
}
\arguments{
\item{data}{A data frame containing plot data: data frame}

\item{x}{Column name representing the continuous data series to be used for x-axis: character}

\item{y}{Column name representing the continuous data series to be used for y-axis: character}

\item{colour}{Column name representing data to be used for colour grouping (Default: NULL): NULL or character}

\item{shape}{Column name representing data to be used for shape grouping (Default: NULL): NULL or character}

\item{title}{Plot title (Default: "Time Series"): NULL or character}

\item{subtitle}{Plot subtitle (Default: NULL): NULL or character}

\item{caption}{Plot caption (Default: NULL): NULL or character}

\item{xLabel}{x-axis label (Default: Value used for x): NULL or character}

\item{yLabel}{y-axis label (Default: Value used for y): NULL or character}

\item{colourLabel}{Colour grouping label (Default: Value used for colour): NULL or character}

\item{shapeLabel}{Shape grouping label (Default: Value used for shape)}

\item{boundary}{Boundary analysis to perform (Default: NULL, Options: "value", "percentage", "sd"): NULL or character.}

\item{boundaryValue}{Value used for "value" and "percentage" boundary analysis (Default: 0): double}

\item{referenceData}{A data frame with the same column names and data to calculate "sd" boundaries (Default: NULL): NULL or data frame}

\item{xTickToggle}{Toggle to display ticks on x-axis (Default: TRUE, Options: TRUE or FALSE): boolean}
}
\value{
A data frame containing the result of the boundary analysis performed
}
\description{
Plot time series data (and perform boundary analysis if defined).
}
\details{
Boundary analysis that can be performed:
\itemize{
\item value: Evaluates data by +/- value from 0
\item percentage: Evaluates data +/- value from 100
\item sd: Evaluates data using 2 standard deviations from average mean of the data or using defined mean and standard deviation values
}

To define mean and standard deviation values to be used for "sd" boundary analysis, pass reference data that mirrors the same columns as data.
}
