% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderTICQ.R
\name{renderTICQ}
\alias{renderTICQ}
\title{Render TICQ Assessment}
\usage{
renderTICQ(
  inputPath,
  outputDirectoryPath = paste0(getwd(), "/output/"),
  historicalDataPath = NULL,
  targetFilePath = NULL,
  anpcMethodLibrary = NULL,
  chromatogramRegion = NULL,
  roundDecimalPlace = NULL,
  metadataAnonymiser = FALSE
)
}
\arguments{
\item{inputPath}{Input path to JSON files generated by ExfilMS: character}

\item{outputDirectoryPath}{Output directory path to save HTML report and spectral data RDS output (Default: /output/folder/in/current/working/directory/): character}

\item{historicalDataPath}{Historical data path to be used for spectral data comparison (Default: NULL, Options: Local file path of a RDS file): NULL or character}

\item{targetFilePath}{Target file path (Default: NULL, Options: Published to web URL or local file path of a TSV file): NULL or character}

\item{anpcMethodLibrary}{ANPC method library (Default: NULL, Options: "MS-AA-POS", "MS-HIL-POS", "MS-HIL-NEG", "MS-RP-POS" or "MS-RP-NEG"): NULL or character}

\item{chromatogramRegion}{A list containing the start and end time points of the different chromatogram regions of interest (Default: NULL, Options: ticq::configureChromatogramRegion()): NULL or list}

\item{roundDecimalPlace}{Number of decimal places for precision value rounding (Default: NULL): NULL or numeric}

\item{metadataAnonymiser}{Toggle to anonymise metadata (Default: FALSE, Options: TRUE or FALSE): logical}
}
\description{
Perform TICQ assessment on MS spectral data acquired from JSON files generated using \href{https://github.com/vmalnathnambiar/exfilms}{ExfilMS}
and render a HTML report of the assessment.
}
\details{
Metadata extraction and anonymisation only applicable to data acquired at the Australian National Phenome Centre (ANPC).

To evaluate extracted ions (target m/z) with TICQ, spectral data generated by ExfilMS should have undergone targeted spectra filtering during the extraction process into JSON using a target file or an in-house ANPC method library.
The same target file or ANPC method library should be specified in TICQ for a comprehensive assessment.

Historical spectral data used for comparison should have also undergone the same extraction/spectra filtering process previously as the current spectral data being assessed.
}
