% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderTICQ.R
\name{renderTICQ}
\alias{renderTICQ}
\title{Render TICQ}
\usage{
renderTICQ(
  inputPath,
  outputDirectoryPath = paste0(getwd(), "/output/"),
  historicalDataPath = NULL,
  targetFilePath = NULL,
  anpcMethodLibrary = NULL,
  chromatogramRegion = NULL,
  roundDecimalPlace = NULL,
  metadataAnonymiser = FALSE
)
}
\arguments{
\item{inputPath}{A character string representing a directory or file path.}

\item{outputDirectoryPath}{A character string representing an output directory path. (Default: "/home/vimalnathnambiar/git/github/vmalnathnambiar/ticq/output/")}

\item{historicalDataPath}{A character string representing a RDS file containing historical MS spectral data for comparison.
(Default: \code{NULL}; Options: \code{"/path/to/local/file.rds"})}

\item{targetFilePath}{A character string representing a TSV target file path.
(Default: \code{NULL}; Options: \code{"https://published-to-web-URL&output=tsv"} or \code{"/path/to/local/file.tsv"})}

\item{anpcMethodLibrary}{A character string representing an ANPC method library.
(Default: \code{NULL}; Options: \code{"MS-AA-POS"}, \code{"MS-HIL-POS"}, \code{"MS-HIL-NEG"}, \code{"MS-RP-POS"}, or \code{"MS-RP-NEG"})}

\item{chromatogramRegion}{A list representing the chromatogram region data of interests. (Default: \code{NULL}; Options: \code{configureChromatogramRegion()})}

\item{roundDecimalPlace}{A numeric value representing the number of decimal places to be used for precision value rounding. (Default: \code{NULL})}

\item{metadataAnonymiser}{A logical value representing the status of the metadata anonymiser to anonymise acquired metadata from ANPC-specific data files.
(Default: \code{FALSE}; Options: \code{TRUE} or \code{FALSE})}
}
\value{
A character string of the rendered HTML TICQ assessment report file path.
}
\description{
Render a HTML report of the TICQ assessment performed on MS spectral data retrieved from
\href{https://github.com/vmalnathnambiar/exfilms}{ExfilMS} extracted JSON files.
}
\details{
\enumerate{
\item Metadata extraction and anonymisation only applicable to data acquired at the Australian National Phenome Centre (ANPC).
\item To perform a targeted assessment of extracted ions (targeted m/z) with TICQ, spectral data extracted using ExfilMS must have undergone
targeted spectra filtering during the extraction process using a target file or an in-house ANPC method library. The same target file or
ANPC method library should then be specified in TICQ for a comprehensive assessment.
\item To perform comparison with a historical spectral data, the data should have also undergone the same extraction process as
the current spectral data being assessed.
}
}
