% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveFileList.R
\name{retrieveFileList}
\alias{retrieveFileList}
\title{Retrieve File List}
\usage{
retrieveFileList(inputPath, fileExtension = NULL)
}
\arguments{
\item{inputPath}{A character string representing an existing directory or file path. Should not be NA or an empty character string.}

\item{fileExtension}{A character string representing a file extension type (e.g., "JSON"). (Default: \code{NULL})}
}
\value{
A list of file names matching the specified file extension type.
}
\description{
Retrieve a list of file names matching a specified file extension type.
}
