% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveFileList.R
\name{retrieveFileList}
\alias{retrieveFileList}
\title{retrieveFileList}
\usage{
retrieveFileList(input, fileExtension = NULL)
}
\arguments{
\item{input}{Input path containing a list of files to retrieve: character}

\item{fileExtension}{File extension to filter for (Default: NULL): NULL or character}
}
\value{
List of files within the input path that matches the file extension specified: character vector
}
\description{
Filter and retrieve file list of a specified extension type from input path.
}
\examples{
inputFiles <- ticq::retrieveFileList(input = "/path/to/input/directory/", fileExtension = "JSON")
}
