% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveFileName.R
\name{retrieveFileName}
\alias{retrieveFileName}
\title{Retrieve File Name}
\usage{
retrieveFileName(inputPath, fileExtension = NULL)
}
\arguments{
\item{inputPath}{A character string representing an existing directory or file path.}

\item{fileExtension}{A character string representing a file extension type (e.g., "JSON"). (Default: \code{NULL})}
}
\value{
A character vector of file names matching the specified file extension type.
}
\description{
Retrieve file names matching a specified file extension type.
}
