% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveFileName.R
\name{retrieveFileName}
\alias{retrieveFileName}
\title{Retrieve File Name}
\usage{
retrieveFileName(inputPath, fileExtension = NULL)
}
\arguments{
\item{inputPath}{Input path: character}

\item{fileExtension}{File extension for file name retrieval (Default: NULL): NULL or character}
}
\value{
File names with matching file extension: character vector
}
\description{
Retrieve file names of a specific extension type stored within an input path.
}
