% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveFilename.R
\name{retrieveFilename}
\alias{retrieveFilename}
\title{retrieveFilename}
\usage{
retrieveFilename(input, fileExtension = NULL)
}
\arguments{
\item{input}{Input path containing files of interest: character}

\item{fileExtension}{File extension to filter for (Default: NULL): NULL or character}
}
\value{
File names matching the file extension specified: character vector
}
\description{
Filter and retrieve names of files of a specific extension type stored within an input path.
}
