% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveFileName.R
\name{retrieveFileName}
\alias{retrieveFileName}
\title{Retrieve File Name}
\usage{
retrieveFileName(inputPath, fileExtension = NULL)
}
\arguments{
\item{inputPath}{A character string representing a directory or file path.}

\item{fileExtension}{A character string representing a file extension pattern (e.g., "JSON"). (Default: \code{NULL})}
}
\value{
A character vector of the file names matching the specified file extension pattern.
}
\description{
Retrieve file names from a directory or file path that matches a specific file extension pattern.
}
