% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceChromatogramRegion.R
\name{spliceChromatogramRegion}
\alias{spliceChromatogramRegion}
\title{Splice Chromatogram Region}
\usage{
spliceChromatogramRegion(
  data,
  commonColumn,
  spectrumCount,
  chromatogramRegion = NULL,
  retentionTime,
  sumColumn
)
}
\arguments{
\item{data}{A data frame containing spectral data.}

\item{commonColumn}{A character vector representing names of the common data columns to be used for data grouping.}

\item{spectrumCount}{A character string representing the name of the spectrum count column.}

\item{chromatogramRegion}{A list of lists representing the different chromatogram regions of interest and their respective start and end time points. (Default: \code{NULL}; Options: Use \code{ticq::configureChromatogramRegion()})}

\item{retentionTime}{A character string representing the name of the retention time column.}

\item{sumColumn}{A character string representing the name of the data column to be summed.}
}
\value{
A data frame containing a summarised listing of samples along with the spliced data of the different chromatogram regions of interest.
}
\description{
Splice spectral data by summing the values of a data column by the respective start and end time points of the different chromatogram regions of interest.
}
