% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceChromatogramRegion.R
\name{spliceChromatogramRegion}
\alias{spliceChromatogramRegion}
\title{spliceChromatogramRegion}
\usage{
spliceChromatogramRegion(
  data,
  commonColumn,
  spectrumCount,
  chromatogramRegion = NULL,
  filterBy,
  sumBy
)
}
\arguments{
\item{data}{Data frame that contains spectral information of samples: data frame}

\item{commonColumn}{Column names that are common across all samples: character vector}

\item{spectrumCount}{Name of the column that represents spectrum count: character}

\item{chromatogramRegion}{List that represents the different chromatogram regions of interest (Default: NULL): NULL or list}

\item{filterBy}{Name of the column that represents the retention time to be filtered: character}

\item{sumBy}{Name of the column that represents the data to be summed for each chromatogram region: character}
}
\value{
Data frame containing the sum values of the spliced chromatogram regions
}
\description{
Splice chromatogram regions and sum values of a specific data column together.
}
