% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceChromatogramRegion.R
\name{spliceChromatogramRegion}
\alias{spliceChromatogramRegion}
\title{Splice Chromatogram Region}
\usage{
spliceChromatogramRegion(
  data,
  commonColumn,
  spectrumCount,
  chromatogramRegion = NULL,
  retentionTime,
  sumColumn
)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{commonColumn}{Column names representing common data to be used for data grouping: character}

\item{spectrumCount}{Spectrum count column name: character}

\item{chromatogramRegion}{A list representing the different chromatogram regions of interest (Default: NULL, Options: ticq::configureChromatogramRegion()): NULL or list}

\item{retentionTime}{Retention time column name: character}

\item{sumColumn}{Column name of the data to be summed: character}
}
\value{
A data frame containing spliced spectral data of the different chromatogram regions of interest
}
\description{
Splice spectral data by summing the values of a column by the respective start and end time points of the different chromatogram regions of interest.
}
