% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceChromatogramRegion.R
\name{spliceChromatogramRegion}
\alias{spliceChromatogramRegion}
\title{Splice Chromatogram Region}
\usage{
spliceChromatogramRegion(
  data,
  commonColumn,
  spectrumCount,
  chromatogramRegion = NULL,
  retentionTime,
  sumColumn
)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{commonColumn}{Column names of data common for each sample: character vector}

\item{spectrumCount}{Spectrum count column name: character}

\item{chromatogramRegion}{A list containing the start and end time points of the different chromatogram regions of interest (prewash, maass calibration, analyte and wash) (Default: NULL): NULL or list}

\item{retentionTime}{Retention time column name: character}

\item{sumColumn}{Column name of the data to be summed: character}
}
\value{
A data frame containing the summed values for each chromatogram regions of interest
}
\description{
Splice chromatogram regions of interest by summing values of a column by the respective start and end time points.
}
\details{
Configure chromatogram region using: ticq::configureChromatogramRegion()
}
