% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceChromatogramRegion.R
\name{spliceChromatogramRegion}
\alias{spliceChromatogramRegion}
\title{spliceChromatogramRegion}
\usage{
spliceChromatogramRegion(
  data,
  commonColumn,
  spectrumCount,
  chromatogramRegion = NULL,
  filterBy,
  sumBy
)
}
\arguments{
\item{data}{Data frame that contains spectral information of samples: data frame}

\item{commonColumn}{A vector of column names that are common across all samples: character vector}

\item{spectrumCount}{The name of the column that represents spectrum count: character}

\item{chromatogramRegion}{A list that represent the different chromatogram regions (Default: NULL): NULL or list}

\item{filterBy}{The name of the column that represents the retention time to be filtered: character}

\item{sumBy}{The name of the column that represents the data to be summed for each chromatogram region: character}
}
\value{
A data frame containing the sum values of the spliced regions of the chromatogram
}
\description{
Splice chromatogram regions and sum values of a specific data column together for each sample.
}
