% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumDataColumn.R
\name{sumDataColumn}
\alias{sumDataColumn}
\title{sumDataColumn}
\usage{
sumDataColumn(data, commonColumn, sumBy)
}
\arguments{
\item{data}{Data frame that contains spectral information of samples: data frame}

\item{commonColumn}{Column names that are common across all samples: character vector}

\item{sumBy}{Name of the column that represents the data to be summed: character}
}
\value{
Data frame grouped by common columns and the sum value of the column specified
}
\description{
Sum values of a data column by grouping using its common columns.
}
