% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumDataColumn.R
\name{sumDataColumn}
\alias{sumDataColumn}
\title{sumDataColumn}
\usage{
sumDataColumn(data, commonColumn, sumBy)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{commonColumn}{Column names of data common for each unique sample: character vector}

\item{sumBy}{Column name of the data to be summed: character}
}
\value{
A data frame grouped by common columns and the sum value of the data column specified
}
\description{
Sum values of a data column by grouping using its common columns.
}
