% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumDataColumn.R
\name{sumDataColumn}
\alias{sumDataColumn}
\title{sumDataColumn}
\usage{
sumDataColumn(data, commonColumn, sumBy)
}
\arguments{
\item{data}{Data frame that contains spectrum information of samples: data frame}

\item{commonColumn}{A vector of column names that are common across all samples: character vector}

\item{sumBy}{The name of the column that represents the data to be summed: character}
}
\value{
A data frame grouped by common columns and the sum value of the column
}
\description{
Sum values of a data column for each sample by grouping data by its common columns.
}
