% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumDataColumn.R
\name{sumDataColumn}
\alias{sumDataColumn}
\title{Sum Data Column}
\usage{
sumDataColumn(data, commonColumn, sumColumn)
}
\arguments{
\item{data}{A data frame containing spectral data.}

\item{commonColumn}{A character vector representing names of the common data columns to be used for data grouping.}

\item{sumColumn}{A character string representing the name of the data column to be summed.}
}
\value{
A data frame grouped by common data columns and the sum value of the data column specified.
}
\description{
Sum the values of a single data column for each data group.
}
