% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumEIC.R
\name{sumEIC}
\alias{sumEIC}
\title{sumEIC}
\usage{
sumEIC(data, commonColumn, spectrumCount, startIDX, endIDX = NULL)
}
\arguments{
\item{data}{A data frame containing spectral data: data frame}

\item{commonColumn}{Column names of data common for each unique sample: character vector}

\item{spectrumCount}{Spectrum count column name: character}

\item{startIDX}{Index of the first data column to summed: double}

\item{endIDX}{Index of the last data column to summed: double}
}
\value{
A data frame grouped by common columns and the summed values of each column specified
}
\description{
Sum EIC values of multiple data columns by grouping using its common columns.
}
