% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumEIC.R
\name{sumEIC}
\alias{sumEIC}
\title{sumEIC}
\usage{
sumEIC(data, commonColumn, spectrumCount, startIDX, endIDX = NULL)
}
\arguments{
\item{data}{Data frame that contains spectral information of samples: data frame}

\item{commonColumn}{Column names that are common across all samples: character vector}

\item{spectrumCount}{Name of the column that represents spectrum count: character}

\item{startIDX}{Index of the first column to sum: double}

\item{endIDX}{Index of the last column to sum: double}
}
\value{
A data frame grouped by common columns and the summed EIC values for each column specified
}
\description{
Sum EIC values by grouping using its common columns.
}
