% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHelper.R
\name{validateCharacterVector}
\alias{validateCharacterVector}
\title{Validate Character Vector}
\usage{
validateCharacterVector(parameterName, parameterValue)
}
\arguments{
\item{parameterName}{A character string representing the name of the parameter being validated.}

\item{parameterValue}{A character vector to be validated.}
}
\value{
Invisible TRUE if the character vector is valid; otherwise, it stops with an error message.
}
\description{
A helper function to validate a character vector.
It ensures that the vector specified has a length greater than 0, is a character type, and does not contain NA or empty strings.
}
