% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{validateChromatogramRegion}
\alias{validateChromatogramRegion}
\title{Validate Chromatogram Region}
\usage{
validateChromatogramRegion(name, value)
}
\arguments{
\item{name}{A character string representing the name of the object being validated.}

\item{value}{A list representing the chromatogram region data of interests.}
}
\value{
Invisible TRUE if the data of the chromatogram regions of interest are valid; otherwise, invisible FALSE along with an error message.
}
\description{
Validate that a list representing the chromatogram region data of interests (prewash, mass calibration, analyte, and wash) has a length of 4;
that each region of interest is also a list and has a length of 2 representing values of its respective time points (start and end);
and that each time point value has a length of 1 and is a numeric type (except 'washEnd' with the acceptance of a NULL value).
}
