% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHelper.R
\name{validateChromatogramRegion}
\alias{validateChromatogramRegion}
\title{Validate Chromatogram Region}
\usage{
validateChromatogramRegion(parameterName, parameterValue)
}
\arguments{
\item{parameterName}{A character string representing the name of the parameter being validated.}

\item{parameterValue}{A list of lists representing the different chromatogram regions of interest and their respective start and end time points.}
}
\value{
Invisible TRUE if all chromatogram region data is valid; otherwise, invisible FALSE along with an error message indicating the cause of invalidity.
}
\description{
A helper function to validate a list containing lists of chromatogram region data (start and end time points).
It ensures that the chromatogram region data specified is a list and has a valid length representing all regions of interests.
The function also ensures the data of each chromatogram region is also stored within a list and has a valid length representing their respective time points,
with each being a numeric value of length 1 (or in some, being NULL).
}
