% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperClass.R
\name{validateDirectoryFileExist}
\alias{validateDirectoryFileExist}
\title{Validate Directory and File Path}
\usage{
validateDirectoryFileExist(parameterName, parameterValue, pathType = "both")
}
\arguments{
\item{parameterName}{A character string representing the name of the parameter being validated.}

\item{parameterValue}{A character string representing either a directory or file path to be validated.}

\item{pathType}{A character string representing the path type to validate for. (Default: \code{"both"}; Options: \code{"directory"}, \code{"file"}, \code{"both"})}
}
\value{
Invisible TRUE (logical) if the parameter is valid; otherwise, it stops with an error message.
}
\description{
A helper function to validate a character string representing either a directory or file path.
It ensures that the path specified has a length of 1, is a character type, is not NA or an empty character string,
and checks if the path exists based on a path type.
}
