% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processHelper.R
\name{validateLogicalValue}
\alias{validateLogicalValue}
\title{Validate Logical Value}
\usage{
validateLogicalValue(parameterName, parameterValue)
}
\arguments{
\item{parameterName}{A character string representing the name of the parameter being validated.}

\item{parameterValue}{A logical value to be validated.}
}
\value{
Invisible TRUE if the logical value is valid; otherwise, it stops with an error message.
}
\description{
A helper function to validate a logical value.
It ensures that the value specified has a length of 1 and is a logical type.
}
